cleaning and separating parts
