+++ HiAnalysis/HiOnia/plugins/HiOniaAnalyzer.cc
@@ -1,619 +1,5 @@
-// -*- C++ -*-
-//
-// Package:    HiOniaAnalyzer
-// Class:      HiOniaAnalyzer
-// 
-/**\class HiOniaAnalyzer HiOniaAnalyzer.cc UserCode/tdahms/HiAnalysis/HiOnia/plugins/HiOniaAnalyzer.cc
+#include "HiAnalysis/HiOnia/interface/HiOniaAnalyzer.h"
 
- Description: [one line class summary]
-
- Implementation:
-     [Notes on implementation]
-*/
-
-// system include files
-#include <memory>
-#include <iostream>
-#include <string>
-#include <sstream>
-#include <vector>
-#include <utility>
-
-#include <TTree.h>
-#include <TLorentzVector.h>
-#include <TVector2.h>
-#include <TClonesArray.h>
-#include <TRegexp.h>
-
-// user include files
-#include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/one/EDAnalyzer.h"
-
-#include "FWCore/Framework/interface/Event.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-
-#include "DataFormats/PatCandidates/interface/Muon.h"
-#include "DataFormats/PatCandidates/interface/CompositeCandidate.h"
-#include "DataFormats/Candidate/interface/CompositeCandidate.h"
-#include <DataFormats/RecoCandidate/interface/RecoChargedCandidate.h>
-#include "DataFormats/Common/interface/TriggerResults.h"
-#include "DataFormats/VertexReco/interface/Vertex.h"
-#include "DataFormats/VertexReco/interface/VertexFwd.h"
-#include "DataFormats/BeamSpot/interface/BeamSpot.h"
-#include "DataFormats/TrackReco/interface/Track.h"
-#include "DataFormats/TrackReco/interface/TrackFwd.h"
-
-#include "HLTrigger/HLTcore/interface/HLTPrescaleProvider.h"
-#include "FWCore/Framework/interface/ConsumesCollector.h"
-#include "HLTrigger/HLTcore/interface/HLTConfigProvider.h"
-
-#include "DataFormats/HeavyIonEvent/interface/Centrality.h"
-#include "DataFormats/HeavyIonEvent/interface/EvtPlane.h"
-#include "SimDataFormats/GeneratorProducts/interface/GenEventInfoProduct.h"
-
-#include "HiAnalysis/HiOnia/interface/MyCommonHistoManager.h"
-#include "DataFormats/Math/interface/deltaR.h"
-
-#include "FWCore/ServiceRegistry/interface/Service.h"
-#include "CommonTools/UtilAlgos/interface/TFileService.h"
-
-//
-// class declaration
-//
-
-class HiOniaAnalyzer : public edm::one::EDAnalyzer<edm::one::SharedResources,edm::one::WatchRuns> {
-public:
-  explicit HiOniaAnalyzer(const edm::ParameterSet&);
-  ~HiOniaAnalyzer();
-  
-  
-private:
-  virtual void beginJob() ;
-  virtual void analyze(const edm::Event&, const edm::EventSetup&);
-  virtual void endJob() ;
-  
-  
-  void InitEvent();
-  void InitTree();
-
-  void makeCuts(bool keepSameSign) ;
-  void makeBcCuts(bool keepWrongSign) ;
-  void makeDimutrkCuts(bool keepWrongSign) ;
-  bool checkCuts(const pat::CompositeCandidate* cand, const pat::Muon* muon1,  const pat::Muon* muon2, bool(HiOniaAnalyzer::* callFunc1)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc2)(const pat::Muon*)); 
-  bool checkBcCuts(const pat::CompositeCandidate* cand, const pat::Muon* muon1,  const pat::Muon* muon2, const pat::Muon* muon3, bool(HiOniaAnalyzer::* callFunc1)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc2)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc3)(const pat::Muon*));
-  bool checkDimuTrkCuts(const pat::CompositeCandidate* cand, const pat::Muon* muon1, const pat::Muon* muon2, const reco::RecoChargedCandidate* trk, bool(HiOniaAnalyzer::* callFunc1)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc2)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc3)(const reco::TrackRef));
-
-  reco::GenParticleRef findDaughterRef(reco::GenParticleRef GenParticleDaughter, int GenParticlePDG);
-  int IndexOfThisMuon(TLorentzVector* v1, bool isGen=false);
-  int IndexOfThisTrack(TLorentzVector* v1, bool isGen=false);
-  int IndexOfThisJpsi(int mu1_idx, int mu2_idx, int flipJpsi=0);
-  void fillGenInfo();
-  void fillMuMatchingInfo();
-  void fillQQMatchingInfo();
-  void fillBcMatchingInfo();
-  bool isAbHadron(int pdgID);
-  bool isNeutrino(int pdgID);
-  bool isAMixedbHadron(int pdgID, int momPdgID);
-  std::pair<bool, reco::GenParticleRef> findBcMotherRef(reco::GenParticleRef GenParticleMother,int GenParticlePDG);
-  bool isChargedTrack(int pdgId);
-  std::vector<reco::GenParticleRef> GenBrothers(reco::GenParticleRef GenParticleMother, int GenJpsiPDG);
-  reco::GenParticleRef findMotherRef(reco::GenParticleRef GenParticleMother, int GenParticlePDG);
-  std::pair<std::vector<reco::GenParticleRef> , std::pair<float, float> >  findGenMCInfo(const reco::GenParticle *genJpsi);
-  std::pair<int, std::pair<float, float> >  findGenBcInfo(reco::GenParticleRef genBc, const reco::GenParticle *genJpsi);
-
-  void fillRecoMuons(int theCentralityBin);
-  bool isMuonInAccept(const pat::Muon* aMuon, std::string muonType);
-  bool isTrkInMuonAccept(TLorentzVector trk4mom, std::string muonType);
-
-  bool isSoftMuonBase(const pat::Muon* aMuon);
-  bool isHybridSoftMuon(const pat::Muon* aMuon);
-  Short_t MuInSV(TLorentzVector v1, TLorentzVector v2, TLorentzVector v3);
-
-  void fillRecoTracks();
-
-  pair< unsigned int, const pat::CompositeCandidate* > theBestQQ();
-  double CorrectMass(const reco::Muon& mu1,const reco::Muon& mu2, int mode);
-
-  bool selGlobalMuon(const pat::Muon* aMuon);
-  bool selTrackerMuon(const pat::Muon* aMuon);
-  bool selGlobalOrTrackerMuon(const pat::Muon* aMuon);
-  bool selAllMuon(const pat::Muon* aMuon);
-  bool selTrk(const reco::TrackRef aTrk);
-  bool PassMiniAODcut(const pat::Muon* aMuon);
-
-  void fillRecoHistos();
-  void fillRecoJpsi(int count, std::string trigName, std::string centName);
-  void fillHistosAndDS(unsigned int theCat, const pat::CompositeCandidate* aJpsiCand);
-
-  void fillTreeMuon(const pat::Muon* muon, int iType, ULong64_t trigBits);
-  void fillTreeJpsi(int count);
-  void fillTreeBc(int count);
-  void fillTreeDimuTrk(int count);
-
-  void checkTriggers(const pat::CompositeCandidate* aJpsiCand);
-  void hltReport(const edm::Event &iEvent ,const edm::EventSetup& iSetup);
-
-  long int FloatToIntkey(float v);
-  void beginRun(const edm::Run &, const edm::EventSetup &); 
-  void endRun(const edm::Run &, const edm::EventSetup &) {};
-
-  TLorentzVector lorentzMomentum(const reco::Candidate::LorentzVector& p);
-  int muonIDmask(const pat::Muon* muon);
-
-  // ----------member data ---------------------------
-  enum StatBins {
-    BIN_nEvents = 0
-  };
-
-  enum dimuonCategories {
-    GlbTrk_GlbTrk     = 0,
-    Glb_Glb           = 1,
-    Trk_Trk           = 2,
-    GlbOrTrk_GlbOrTrk = 3,
-    TwoGlbAmongThree  = 4,
-    All_All           = 5
-  };
-
-  enum muonCategories {
-    GlbTrk        = 0,
-    Trk           = 1,
-    Glb           = 2,
-    GlbOrTrk      = 3,
-    All           = 4
-  };
-
-  std::vector<std::string> theRegions;
-  std::vector<std::string> theCentralities;
-  std::vector<std::string> theTriggerNames;
-  std::vector<std::string> theSign;
-  std::map<std::string, std::string> triggerNameMap;
-  std::map<std::string, std::string> filterNameMap;
-
-  HLTConfigProvider hltConfig;
-  bool hltConfigInit;
-
-  float etaMin;
-  float etaMax;
-
-  // TFileService
-  edm::Service<TFileService> fs;
-
-  // // TFile
-  // TFile* fOut;
-
-  // TTree
-  TTree* myTree;
-
-  TClonesArray* Reco_mu_4mom;
-  TClonesArray* Reco_mu_L1_4mom;
-  TClonesArray* Reco_QQ_4mom;
-  TClonesArray* Reco_QQ_mumi_4mom;
-  TClonesArray* Reco_QQ_mupl_4mom;
-  TClonesArray* Reco_3mu_4mom;
-  TClonesArray* Reco_QQ_vtx;
-  TClonesArray* Reco_3mu_vtx;
-  TClonesArray* Reco_trk_4mom;
-  TClonesArray* Reco_trk_vtx;
-  TClonesArray* Gen_Bc_4mom;
-  TClonesArray* Gen_Bc_nuW_4mom;
-  TClonesArray* Gen_3mu_4mom;
-
-  TClonesArray* Gen_mu_4mom;
-  TClonesArray* Gen_QQ_4mom;
-
-  std::vector<float> Reco_mu_4mom_pt;
-  std::vector<float> Reco_mu_L1_4mom_pt;
-  std::vector<float> Reco_QQ_4mom_pt;
-  std::vector<float> Reco_QQ_mumi_4mom_pt;
-  std::vector<float> Reco_QQ_mupl_4mom_pt;
-  std::vector<float> Reco_3mu_4mom_pt;
-  std::vector<float> Reco_trk_4mom_pt;
-  std::vector<float> Gen_Bc_4mom_pt;
-  std::vector<float> Gen_Bc_nuW_4mom_pt;
-  std::vector<float> Gen_3mu_4mom_pt;
-
-  std::vector<float> Gen_mu_4mom_pt;
-  std::vector<float> Gen_QQ_4mom_pt;
-
-  std::vector<float> Reco_mu_4mom_eta;
-  std::vector<float> Reco_mu_L1_4mom_eta;
-  std::vector<float> Reco_QQ_4mom_eta;
-  std::vector<float> Reco_QQ_mumi_4mom_eta;
-  std::vector<float> Reco_QQ_mupl_4mom_eta;
-  std::vector<float> Reco_3mu_4mom_eta;
-  std::vector<float> Reco_trk_4mom_eta;
-  std::vector<float> Gen_Bc_4mom_eta;
-  std::vector<float> Gen_Bc_nuW_4mom_eta;
-  std::vector<float> Gen_3mu_4mom_eta;
-
-  std::vector<float> Gen_mu_4mom_eta;
-  std::vector<float> Gen_QQ_4mom_eta;
-
-  std::vector<float> Reco_mu_4mom_phi;
-  std::vector<float> Reco_mu_L1_4mom_phi;
-  std::vector<float> Reco_QQ_4mom_phi;
-  std::vector<float> Reco_QQ_mumi_4mom_phi;
-  std::vector<float> Reco_QQ_mupl_4mom_phi;
-  std::vector<float> Reco_3mu_4mom_phi;
-  std::vector<float> Reco_trk_4mom_phi;
-  std::vector<float> Gen_Bc_4mom_phi;
-  std::vector<float> Gen_Bc_nuW_4mom_phi;
-  std::vector<float> Gen_3mu_4mom_phi;
-
-  std::vector<float> Gen_mu_4mom_phi;
-  std::vector<float> Gen_QQ_4mom_phi;
-
-  std::vector<float> Reco_mu_4mom_m;
-  std::vector<float> Reco_mu_L1_4mom_m;
-  std::vector<float> Reco_QQ_4mom_m;
-  std::vector<float> Reco_QQ_mumi_4mom_m;
-  std::vector<float> Reco_QQ_mupl_4mom_m;
-  std::vector<float> Reco_3mu_4mom_m;
-  std::vector<float> Reco_trk_4mom_m;
-  std::vector<float> Gen_Bc_4mom_m;
-  std::vector<float> Gen_Bc_nuW_4mom_m;
-  std::vector<float> Gen_3mu_4mom_m;
-
-  std::vector<float> Gen_mu_4mom_m;
-  std::vector<float> Gen_QQ_4mom_m;
-
-  static const int Max_QQ_size = 10000;
-  static const int Max_Bc_size = 10000;
-  static const int Max_mu_size = 1000;
-  static const int Max_trk_size = 10000;
-
-  float Gen_weight; // generator weight
-  float Gen_pthat; // ptHat scale of generated hard scattering
-
-  Short_t Gen_QQ_size; // number of generated Onia
-  Short_t Gen_QQ_type[Max_QQ_size]; // Onia type: prompt, non-prompt, unmatched
-  float Gen_QQ_ctau[Max_QQ_size];    // ctau: flight time
-  float Gen_QQ_ctau3D[Max_QQ_size];    // ctau3D: 3D flight time
-  int Gen_QQ_momId[Max_QQ_size]; // PDG ID of the generated mother of the Gen QQ, going back far enough in the geneaology to find a potential B mother
-  float Gen_QQ_momPt[Max_QQ_size];    // Pt of mother particle of 2 muons
-  Short_t Gen_QQ_mupl_idx[Max_QQ_size];    // index of the muon plus from Jpsi, in the full list of muons
-  Short_t Gen_QQ_mumi_idx[Max_QQ_size];    // index of the muon minus from Jpsi, in the full list of muons
-  Short_t Gen_QQ_whichRec[Max_QQ_size]; // index of the reconstructed Jpsi that was matched with this gen Jpsi. Is -1 if one of the 2 muons from Jpsi was not reconstructed. Is -2 if the two muons were reconstructed, but the dimuon was not selected
-  Short_t Gen_QQ_Bc_idx[Max_QQ_size]; //Index of the Bc gen mother. -1 if there is no Bc mother
-
-  Short_t Gen_Bc_size; // number of generated Bc
-  float Gen_Bc_ctau[Max_Bc_size];    // ctau: flight time
-  int Gen_Bc_pdgId[Max_Bc_size];    // pdgId of Bc
-  Short_t Gen_Bc_QQ_idx[Max_Bc_size]; //Points to the number of the associated Jpsi daughter 
-  Short_t Gen_Bc_muW_idx[Max_Bc_size];    // index of the muon from W from Bc, in the full list of muons
-  Short_t Gen_3mu_whichRec[Max_Bc_size]; // index of the reconstructed trimuon that was matched with this gen visible Bc. Is -1 if one muon of the Bc was not reconstructed
- 
-  Short_t Gen_mu_size; // number of generated muons
-  Short_t Gen_mu_charge[Max_mu_size]; // muon charge
-  Short_t Gen_mu_type[Max_mu_size]; // muon type: prompt, non-prompt, unmatched
-  Short_t Gen_mu_whichRec[Max_mu_size]; // index of the reconstructed muon that was matched with this gen muon. Is -1 if the muon was not reconstructed
-  float Gen_mu_MatchDeltaR[Max_mu_size]; // deltaR between reco and gen matched muons
-
-  Short_t Reco_3mu_size;       // Number of reconstructed trimuons
-  Short_t Reco_3mu_charge[Max_Bc_size];
-  Short_t Reco_3mu_mupl_idx[Max_Bc_size];    // index of the muon plus from Jpsi, in the full list of muons
-  Short_t Reco_3mu_mumi_idx[Max_Bc_size];    // index of the muon minus from Jpsi, in the full list of muons
-  Short_t Reco_3mu_muW_idx[Max_Bc_size];    // index of the muon from W, in the full list of muons
-  Short_t Reco_3mu_muW2_idx[Max_Bc_size];    // index of the muon from W, for the second possible OS dimuon
-  //  Short_t Reco_3mu_muW_trkIdx[Max_Bc_size];    // index of the generated track closest to the reco muon from W, when the Jpsi is true and the muon is fake
-  Short_t Reco_3mu_QQ1_idx[Max_Bc_size];    // index of a possible Jpsi from Bc, 1st of the two OS muon pairs
-  Short_t Reco_3mu_QQ2_idx[Max_Bc_size];    // index of a possible Jpsi from Bc, 2nd of the two OS muon pairs
-  Short_t Reco_3mu_QQss_idx[Max_Bc_size];    // index of a the same-sign muon pair from Bc
-  Short_t Reco_3mu_whichGen[Max_Bc_size]; // index of the generated Bc that was matched with this rec Bc. Is -1 if one or more of the 3 muons from Bc was not reconstructed
-  bool Reco_3mu_muW_isGenJpsiBro[Max_Bc_size]; // is this true or fake muon matched to a generated particle that is brother/nephew to the gen Jpsi (daughter of the gen B)
-  int Reco_3mu_muW_trueId[Max_Bc_size]; //pdgId of the generated particle that the reco_muW is matched to
-  float Reco_3mu_VtxProb[Max_Bc_size]; // chi2 probability of vertex fitting 
-  float Reco_3mu_KCVtxProb[Max_Bc_size]; // chi2 probability of kinematic constrained vertex fitting 
-  float Reco_3mu_ctau[Max_Bc_size];    // ctau: flight time
-  float Reco_3mu_ctauErr[Max_Bc_size]; // error on ctau
-  float Reco_3mu_cosAlpha[Max_QQ_size];    // cosine of angle between momentum of Bc and direction of PV--displaced vertex segment (in XY plane)
-  float Reco_3mu_ctau3D[Max_Bc_size];    // ctau: flight time in 3D
-  float Reco_3mu_ctauErr3D[Max_Bc_size]; // error on ctau in 3D
-  float Reco_3mu_cosAlpha3D[Max_QQ_size];    // cosine of angle between momentum of Bc and direction of PV--displaced vertex segment (3D)
-  float Reco_3mu_KCctau[Max_Bc_size];    // ctau: flight time
-  float Reco_3mu_KCctauErr[Max_Bc_size]; // error on ctau
-  float Reco_3mu_KCcosAlpha[Max_QQ_size];    // cosine of angle between momentum of Bc and direction of PV--displaced vertex segment (in XY plane)
-  float Reco_3mu_KCctau3D[Max_Bc_size];    // ctau: flight time in 3D
-  float Reco_3mu_KCctauErr3D[Max_Bc_size]; // error on ctau in 3D
-  float Reco_3mu_KCcosAlpha3D[Max_QQ_size];    // cosine of angle between momentum of Bc and direction of PV--displaced vertex segment (3D)
-  float Reco_3mu_MassErr[Max_Bc_size];
-  float Reco_3mu_CorrM[Max_Bc_size];
-  Short_t Reco_3mu_NbMuInSameSV[Max_Bc_size];
-  float Reco_3mu_muW_dxy[Max_Bc_size];
-  float Reco_3mu_muW_dz[Max_Bc_size];
-  float Reco_3mu_mumi_dxy[Max_Bc_size];
-  float Reco_3mu_mumi_dz[Max_Bc_size];
-  float Reco_3mu_mupl_dxy[Max_Bc_size];
-  float Reco_3mu_mupl_dz[Max_Bc_size];
-
-  Short_t Reco_QQ_size;       // Number of reconstructed Onia 
-  Short_t Reco_QQ_type[Max_QQ_size];   // Onia category: GG, GT, TT
-  Short_t Reco_QQ_sign[Max_QQ_size];   /* Mu Mu combinations sign:
-                             0 = +/- (signal)
-                             1 = +/+
-                             2 = -/- 
-                          */
-  Short_t Reco_QQ_mupl_idx[Max_QQ_size];    // index of the muon plus from Jpsi, in the full list of muons
-  Short_t Reco_QQ_mumi_idx[Max_QQ_size];    // index of the muon minus from Jpsi, in the full list of muons
-  Short_t Reco_QQ_whichGen[Max_QQ_size]; // index of the generated Jpsi that was matched with this rec Jpsi. Is -1 if one of the 2 muons from Jpsi was not reconstructed
-  ULong64_t Reco_QQ_trig[Max_QQ_size];      // Vector of trigger bits matched to the Onia
-  float Reco_QQ_VtxProb[Max_QQ_size]; // chi2 probability of vertex fitting 
-  float Reco_QQ_ctau[Max_QQ_size];    // ctau: flight time
-  float Reco_QQ_ctauErr[Max_QQ_size]; // error on ctau
-  float Reco_QQ_cosAlpha[Max_QQ_size];    // cosine of angle between momentum of Jpsi and direction of PV--displaced vertex segment (in XY plane)
-  float Reco_QQ_ctau3D[Max_QQ_size];    // ctau: flight time in 3D
-  float Reco_QQ_ctauErr3D[Max_QQ_size]; // error on ctau in 3D
-  float Reco_QQ_cosAlpha3D[Max_QQ_size];    // cosine of angle between momentum of Jpsi and direction of PV--displaced vertex segment (3D)
-  float Reco_QQ_dca[Max_QQ_size];
-  float Reco_QQ_MassErr[Max_QQ_size];
-
-  int  Reco_QQ_NtrkPt02[Max_QQ_size];
-  int  Reco_QQ_NtrkPt03[Max_QQ_size];
-  int  Reco_QQ_NtrkPt04[Max_QQ_size];
-
-  int  Reco_QQ_NtrkDeltaR03[Max_QQ_size];
-  int  Reco_QQ_NtrkDeltaR04[Max_QQ_size];
-  int  Reco_QQ_NtrkDeltaR05[Max_QQ_size];
-
-  float Reco_QQ_mupl_dxy[Max_QQ_size];  // dxy for plus inner track muons
-  float Reco_QQ_mumi_dxy[Max_QQ_size];  // dxy for minus inner track muons
-  float Reco_QQ_mupl_dz[Max_QQ_size];  // dz for plus inner track muons
-  float Reco_QQ_mumi_dz[Max_QQ_size];  // dz for minus inner track muons
-  Short_t Reco_QQ_flipJpsi[Max_QQ_size];
-
-  Short_t Reco_mu_size;           // Number of reconstructed muons
-  int Reco_mu_SelectionType[Max_mu_size];           
-  ULong64_t Reco_mu_trig[Max_mu_size];      // Vector of trigger bits matched to the muon
-  Short_t Reco_mu_charge[Max_mu_size];  // Vector of charge of muons
-  Short_t Reco_mu_type[Max_mu_size];  // Vector of type of muon (global=0, tracker=1, calo=2)  
-  Short_t Reco_mu_whichGen[Max_mu_size]; // index of the generated muon that was matched with this reco muon. Is -1 if the muon is not associated with a generated muon (fake, or very bad resolution)
-
-  bool Reco_mu_highPurity[Max_mu_size];    // Vector of high purity flag  
-  bool Reco_mu_TrkMuArb[Max_mu_size];      // Vector of TrackerMuonArbitrated
-  bool Reco_mu_TMOneStaTight[Max_mu_size]; // Vector of TMOneStationTight
-  Short_t Reco_mu_candType[Max_mu_size]; // candidate type of muon. 0 (or not present): muon collection, 1: packedPFCandidate, 2: lostTrack collection
-  bool Reco_mu_isPF[Max_mu_size];           // Vector of isParticleFlow muon
-  bool Reco_mu_isTracker[Max_mu_size];
-  bool Reco_mu_isGlobal[Max_mu_size];
-  bool Reco_mu_isSoftCutBased[Max_mu_size];
-  bool Reco_mu_isHybridSoft[Max_mu_size];
-  bool Reco_mu_isMedium[Max_mu_size];
-  bool Reco_mu_isTightCutBased[Max_mu_size];
-  bool Reco_mu_InTightAcc[Max_mu_size];  // Is in the tight acceptance for global muons
-  bool Reco_mu_InLooseAcc[Max_mu_size];  // Is in the loose acceptance for global muons
-
-  int Reco_mu_nPixValHits[Max_mu_size];  // Number of valid pixel hits in sta muons
-  int Reco_mu_nMuValHits[Max_mu_size];  // Number of valid muon hits in sta muons
-  int Reco_mu_nTrkHits[Max_mu_size];  // track hits global muons
-  int Reco_mu_nPixWMea[Max_mu_size];  // pixel layers with measurement for inner track muons
-  int Reco_mu_nTrkWMea[Max_mu_size];  // track layers with measurement for inner track muons
-  int Reco_mu_StationsMatched[Max_mu_size];  // number of stations matched for inner track muons
-  float Reco_mu_segmentComp[Max_mu_size];
-  float Reco_mu_kink[Max_mu_size];
-  float Reco_mu_localChi2[Max_mu_size];
-  float Reco_mu_normChi2_bestTracker[Max_mu_size];
-  float Reco_mu_normChi2_inner[Max_mu_size];  // chi2/ndof for inner track muons
-  float Reco_mu_normChi2_global[Max_mu_size];  // chi2/ndof for global muons
-  float Reco_mu_dxy[Max_mu_size];  // dxy for inner track muons
-  float Reco_mu_dxyErr[Max_mu_size];  // dxy error for inner track muons
-  float Reco_mu_dz[Max_mu_size];  // dz for inner track muons
-  float Reco_mu_dzErr[Max_mu_size];  // dz error for inner track muons
-  float Reco_mu_pt_inner[Max_mu_size];  // pT for inner track muons
-  float Reco_mu_pt_global[Max_mu_size];  // pT for global muons
-  float Reco_mu_ptErr_inner[Max_mu_size];  // pT error for inner track muons
-  float Reco_mu_ptErr_global[Max_mu_size];  // pT error for global muons
-  float Reco_mu_pTrue[Max_mu_size];  // P of the associated generated muon, used to match the Reco_mu with the Gen_mu
-  float Reco_mu_validFraction[Max_mu_size];
-  int Reco_mu_simExtType[Max_Bc_size]; //
-
-  Short_t muType; // type of muon (GlbTrk=0, Trk=1, Glb=2, none=-1) 
-  std::vector<float> EtaOfWantedMuons; //To know which single muons to fill, when fillSingleMuons=false and we want only the muons from selected dimuons
-  std::vector<float> EtaOfWantedTracks; //To know which tracks to fill (only tracks that are part of a selected dimuon-track candidate)
-
-  Short_t Reco_trk_size;           // Number of reconstructed tracks
-  Short_t Reco_trk_charge[Max_trk_size];  // Vector of charge of tracks
-  Short_t Reco_trk_whichGenmu[Max_trk_size];
-  bool Reco_trk_InLooseAcc[Max_trk_size];
-  bool Reco_trk_InTightAcc[Max_trk_size];
-  float Reco_trk_dxyError[Max_trk_size];
-  float Reco_trk_dzError[Max_trk_size];
-  float Reco_trk_dxy[Max_trk_size];
-  float Reco_trk_dz[Max_trk_size];
-  float Reco_trk_ptErr[Max_trk_size];
-  int Reco_trk_originalAlgo[Max_trk_size];
-  int Reco_trk_nPixWMea[Max_trk_size];
-  int Reco_trk_nTrkWMea[Max_trk_size];
-
-  // histos
-  TH1F* hGoodMuonsNoTrig = NULL;
-  TH1F* hGoodMuons = NULL;
-  TH1F* hL1DoubleMu0 = NULL;
-
-  MyCommonHistoManager* myRecoMuonHistos = NULL;
-  MyCommonHistoManager* myRecoGlbMuonHistos = NULL;
-  MyCommonHistoManager* myRecoTrkMuonHistos = NULL;
-
-  MyCommonHistoManager* myRecoJpsiHistos = NULL;
-  MyCommonHistoManager* myRecoJpsiGlbGlbHistos = NULL;
-  MyCommonHistoManager* myRecoJpsiGlbTrkHistos = NULL;
-  MyCommonHistoManager* myRecoJpsiTrkTrkHistos = NULL;
-
-  // event counters
-  TH1F* hStats = NULL;
-
-  // centrality
-  TH1F *hCent = NULL;
-
-  // number of primary vertices
-  TH1F* hPileUp = NULL;
-
-  // z vertex distribution
-  TH1F* hZVtx = NULL;
-
-  // centrality
-  int centBin;
-  int theCentralityBin;
-
-  Short_t Npix, NpixelTracks, Ntracks;
-  int NtracksPtCut, NtracksEtaCut, NtracksEtaPtCut;
-  float SumET_HF, SumET_HFplus, SumET_HFminus, SumET_HFplusEta4, SumET_HFminusEta4,SumET_HFhit, SumET_HFhitPlus, SumET_HFhitMinus, SumET_EB, SumET_ET, SumET_EE, SumET_EEplus, SumET_EEminus, SumET_ZDC, SumET_ZDCplus, SumET_ZDCminus;
-
-  // Event Plane variables
-  int nEP;   // number of event planes
-  //float *hiEvtPlane;
-  float rpAng[50];
-  float rpCos[50];
-  float rpSin[50];
-
-  // handles
-  edm::Handle<pat::CompositeCandidateCollection> collJpsi;
-  edm::Handle<pat::CompositeCandidateCollection> collTrimuon;
-  edm::Handle<pat::CompositeCandidateCollection> collDimutrk;
-  edm::Handle<pat::MuonCollection> collMuon;
-  edm::Handle<pat::MuonCollection> collMuonNoTrig;
-  edm::Handle<reco::TrackCollection> collTracks;
-  edm::Handle<reco::VertexCollection> SVs;
-
-  edm::Handle<reco::GenParticleCollection> collGenParticles;
-  edm::Handle<GenEventInfoProduct> genInfo;
-
-  edm::Handle<edm::TriggerResults> collTriggerResults;
-
-  // data members
-  edm::EDGetTokenT<pat::MuonCollection>               _patMuonToken;
-  edm::EDGetTokenT<pat::MuonCollection>               _patMuonNoTrigToken;
-  edm::EDGetTokenT<pat::CompositeCandidateCollection> _patJpsiToken;
-  edm::EDGetTokenT<pat::CompositeCandidateCollection> _patTrimuonToken;
-  edm::EDGetTokenT<pat::CompositeCandidateCollection> _patDimuTrkToken;
-  edm::EDGetTokenT<reco::TrackCollection>             _recoTracksToken;
-  edm::EDGetTokenT<reco::GenParticleCollection>       _genParticleToken;
-  edm::EDGetTokenT<GenEventInfoProduct>               _genInfoToken;
-  edm::EDGetTokenT<reco::VertexCollection>            _thePVsToken;
-  edm::EDGetTokenT<reco::VertexCollection>            _SVToken;
-  edm::EDGetTokenT<edm::TriggerResults>               _tagTriggerResultsToken;
-  edm::EDGetTokenT<reco::Centrality>                  _centralityTagToken;
-  edm::EDGetTokenT<int>                               _centralityBinTagToken;
-  edm::EDGetTokenT<reco::EvtPlaneCollection>          _evtPlaneTagToken;
-  std::string         _histfilename;
-  std::string         _datasetname;
-  std::string         _mom4format;
-  std::string         _muonSel;
- 
-  std::vector<double> _centralityranges;
-  std::vector<double> _ptbinranges;
-  std::vector<double> _etabinranges;
-  std::vector<string> _dblTriggerPathNames;
-  std::vector<string> _sglTriggerPathNames;
-
-  bool           _onlythebest;
-  bool           _applycuts;
-  bool           _SofterSgMuAcceptance;
-  bool           _SumETvariables;
-  bool           _selTightGlobalMuon;
-  bool           _storeefficiency;
-  bool           _muonLessPrimaryVertex;
-  bool           _useSVfinder;
-  bool           _useBS;
-  bool           _useRapidity;
-  bool           _removeSignal;
-  bool           _removeMuons;
-  bool           _storeSs;
-  bool           _AtLeastOneCand;
-  bool           _combineCategories;
-  bool           _fillRooDataSet;
-  bool           _fillTree;
-  bool           _fillHistos;
-  bool           _theMinimumFlag;
-  bool           _fillSingleMuons;
-  bool           _onlySingleMuons;
-  bool           _fillRecoTracks;
-  bool           _isHI;
-  bool           _isPA;
-  bool           _isMC;
-  bool           _isPromptMC;
-  bool           _useEvtPlane;
-  bool           _useGeTracks;
-  bool           _doTrimuons;
-  bool           _doDimuTrk;
-  int            _flipJpsiDirection;
-  bool           _genealogyInfo;
-  bool           _miniAODcut;
-
-  int _oniaPDG;
-  int _BcPDG;
-  int _OneMatchedHLTMu;
-  bool           _checkTrigNames;
-
-  std::vector<unsigned int>                     _thePassedCats;
-  std::vector<const pat::CompositeCandidate*>   _thePassedCands;
-  std::vector<unsigned int>                     _thePassedBcCats;
-  std::vector<const pat::CompositeCandidate*>   _thePassedBcCands;
-
-  std::vector<reco::GenParticleRef> _Gen_QQ_MomAndTrkBro[Max_QQ_size];
-
-  // number of events
-  unsigned int nEvents;
-  unsigned int passedCandidates;
-
-  unsigned int runNb;
-  unsigned int eventNb;
-  unsigned int lumiSection;
-
-  // limits 
-  float JpsiMassMin;
-  float JpsiMassMax;
-  float JpsiPtMin;           // SET BY 
-  float JpsiPtMax;           // DEFINITION
-  float JpsiRapMin;          // OF BIN
-  float JpsiRapMax;          // LIMITS 
-  float JpsiPDGMass = 3.096916;
-  float BcPDGMass = 6.276;
-
-  math::XYZPoint RefVtx;
-  float RefVtx_xError;
-  float RefVtx_yError;
-  float RefVtx_zError;
-  float zVtx;
-  Short_t nPV;
-
-  // Trigger stuff
-  // PUT HERE THE *LAST FILTERS* OF THE BITS YOU LIKE
-  static const unsigned int sNTRIGGERS = 65;
-  unsigned int NTRIGGERS;
-  unsigned int NTRIGGERS_DBL;
-  unsigned int nTrig;
-
-  // MC 8E29
-  bool isTriggerMatched[sNTRIGGERS];
-  bool alreadyFilled[sNTRIGGERS];
-  ULong64_t HLTriggers;
-  int trigPrescale[sNTRIGGERS];
-
-  std::map<std::string, int> mapTriggerNameToIntFired_;
-  std::map<std::string, int> mapTriggerNameToPrescaleFac_;
-  std::map<long int, int> mapMuonMomToIndex_;
-  std::map<long int, int> mapGenMuonMomToIndex_;
-  std::map<long int, int> mapTrkMomToIndex_;
-
-  HLTPrescaleProvider hltPrescaleProvider;
-  bool hltPrescaleInit;
-
-  const edm::ParameterSet _iConfig;
-};
-
-//
-// constants, enums and typedefs
-//
-
-//
-// static data member definitions
-//
-
-//
-// constructors and destructor
-//
 HiOniaAnalyzer::HiOniaAnalyzer(const edm::ParameterSet& iConfig):
   _patMuonToken(consumes<pat::MuonCollection>(iConfig.getParameter<edm::InputTag>("srcMuon"))),
   _patMuonNoTrigToken(consumes<pat::MuonCollection>(iConfig.getParameter<edm::InputTag>("srcMuonNoTrig"))),
@@ -676,8 +62,7 @@ HiOniaAnalyzer::HiOniaAnalyzer(const edm::ParameterSet& iConfig):
   _OneMatchedHLTMu(iConfig.getParameter<int>("OneMatchedHLTMu")),
   _checkTrigNames(iConfig.getParameter<bool>("checkTrigNames")),
   hltPrescaleProvider(iConfig, consumesCollector(), *this),
-  _iConfig(iConfig)
-{
+  _iConfig(iConfig){
   usesResource(TFileService::kSharedResource);
 
   if(_doTrimuons && _doDimuTrk){
@@ -763,11 +148,10 @@ HiOniaAnalyzer::HiOniaAnalyzer(const edm::ParameterSet& iConfig):
       mapTriggerNameToIntFired_[*it] = -9999;
       mapTriggerNameToPrescaleFac_[*it] = -1;
   }
-}
+};
 
 
-HiOniaAnalyzer::~HiOniaAnalyzer()
-{
+HiOniaAnalyzer::~HiOniaAnalyzer(){
  
   // do anything here that needs to be done at destruction time
   // (e.g. close files, deallocate resources etc.)
@@ -794,17 +178,10 @@ HiOniaAnalyzer::~HiOniaAnalyzer()
     Gen_mu_4mom->Delete();
     Gen_QQ_4mom->Delete();
   }
-}
+};
 
 
-//
-// member functions
-//
-
-// ------------ method called to for each event  ------------
-void
-HiOniaAnalyzer::analyze(const edm::Event& iEvent, const edm::EventSetup& iSetup)
-{
+void HiOniaAnalyzer::analyze(const edm::Event& iEvent, const edm::EventSetup& iSetup){
   //   using namespace edm;
   InitEvent();
   nEvents++;
@@ -1026,10 +403,9 @@ HiOniaAnalyzer::analyze(const edm::Event& iEvent, const edm::EventSetup& iSetup)
     myTree->Fill();
   
   return;
-}
+};
 
-void
-HiOniaAnalyzer::fillRecoHistos() {
+void HiOniaAnalyzer::fillRecoHistos(){
 
   if(!_doTrimuons || !_isMC || _thePassedBcCands.size()>0){ //not storing the mu and QQ reconstructed info if we do a trimuon MC and there is no reco trimuon
     // BEST J/PSI? 
@@ -1070,10 +446,9 @@ HiOniaAnalyzer::fillRecoHistos() {
   }
 
   return;
-}
+};
 
-void
-HiOniaAnalyzer::fillTreeMuon(const pat::Muon* muon, int iType, ULong64_t trigBits) {
+void HiOniaAnalyzer::fillTreeMuon(const pat::Muon* muon, int iType, ULong64_t trigBits){
   if (Reco_mu_size >= Max_mu_size) {
     std::cout << "Too many muons: " << Reco_mu_size << std::endl;
     std::cout << "Maximum allowed: " << Max_mu_size << std::endl;
@@ -1185,10 +560,9 @@ HiOniaAnalyzer::fillTreeMuon(const pat::Muon* muon, int iType, ULong64_t trigBit
 
   Reco_mu_size++;
   return;
-}
+};
 
-void
-HiOniaAnalyzer::fillTreeJpsi(int count) {
+void HiOniaAnalyzer::fillTreeJpsi(int count){
   if (Reco_QQ_size >= Max_QQ_size) {
     std::cout << "Too many dimuons: " << Reco_QQ_size << std::endl;
     std::cout << "Maximum allowed: " << Max_QQ_size << std::endl;
@@ -1468,510 +842,9 @@ HiOniaAnalyzer::fillTreeJpsi(int count) {
 
   Reco_QQ_size++;
   return;
-}
+};
 
-void
-HiOniaAnalyzer::fillTreeBc(int count) {
-
-  if (Reco_3mu_size >= Max_Bc_size) {
-    std::cout << "Too many trimuons: " << Reco_3mu_size << std::endl;
-    std::cout << "Maximum allowed: " << Max_Bc_size << std::endl;
-    return;
-  }
-
-  const pat::CompositeCandidate* aBcCand = _thePassedBcCands.at(count);
-  
-  if (aBcCand==NULL){
-    std::cout<<"ERROR: 'aBcCand' pointer in fillTreeBc is NULL ! Return now"<<std::endl; return;
-  } else {
-    const pat::Muon* muon1 = dynamic_cast<const pat::Muon*>(aBcCand->daughter("muon1"));
-    const pat::Muon* muon2 = dynamic_cast<const pat::Muon*>(aBcCand->daughter("muon2"));
-    const pat::Muon* muon3 = dynamic_cast<const pat::Muon*>(aBcCand->daughter("muon3"));
-    
-    if (muon1==NULL || muon2==NULL || muon3==NULL){
-      std::cout<<"ERROR: 'muon1' or 'muon2' or 'muon3' pointer in fillTreeBc is NULL ! Return now"<<std::endl; return;
-    } else {
-      
-      int charge = muon1->charge() + muon2->charge() + muon3->charge(); 
-
-      TLorentzVector vMuon1 = lorentzMomentum(muon1->p4());
-      TLorentzVector vMuon2 = lorentzMomentum(muon2->p4());
-      TLorentzVector vMuon3 = lorentzMomentum(muon3->p4());
-
-      int mu1_idx = IndexOfThisMuon(&vMuon1); //the muon list contains unchanged muons (even in jpsiFlipping case)
-      int mu2_idx = IndexOfThisMuon(&vMuon2);
-      int mu3_idx = IndexOfThisMuon(&vMuon3);
-
-      int flipJ = (aBcCand->hasUserInt("flipJpsi")) ? aBcCand->userInt("flipJpsi") : 0;
-
-      //One dimuon combination has to pass the Jpsi kinematic cuts
-      if (IndexOfThisJpsi(mu1_idx,mu2_idx,flipJ)==-1 && IndexOfThisJpsi(mu2_idx,mu3_idx,flipJ)==-1 && IndexOfThisJpsi(mu1_idx,mu3_idx,flipJ)==-1) {return;}
-      
-      if(_flipJpsiDirection>0){ //in case of Jpsi flipping, we know that mu1-mu2 is the chosen Jpsi OS dimuon
-	Reco_3mu_QQ1_idx[Reco_3mu_size] = IndexOfThisJpsi(mu1_idx,mu2_idx,flipJ);
-	Reco_3mu_muW_idx[Reco_3mu_size] = mu3_idx;
-	Reco_3mu_mumi_idx[Reco_3mu_size] = (Reco_mu_charge[mu1_idx]==-1)?mu1_idx:mu2_idx;
-	Reco_3mu_mupl_idx[Reco_3mu_size] = (Reco_mu_charge[mu1_idx]==1)?mu1_idx:mu2_idx;
-        
-	if(Reco_mu_charge[mu1_idx]==Reco_mu_charge[mu3_idx]){  //supposing mu1 and mu2 are precedently chosen as opposite-sign
-	  Reco_3mu_QQ2_idx[Reco_3mu_size] = IndexOfThisJpsi(mu3_idx,mu2_idx,flipJ);
-	  Reco_3mu_QQss_idx[Reco_3mu_size] = IndexOfThisJpsi(mu3_idx,mu1_idx,flipJ);
-	} else {
-	  Reco_3mu_QQ2_idx[Reco_3mu_size] = IndexOfThisJpsi(mu3_idx,mu1_idx,flipJ);
-	  Reco_3mu_QQss_idx[Reco_3mu_size] = IndexOfThisJpsi(mu3_idx,mu2_idx,flipJ);
-	}
-      }// end jpsi flipping case
-
-      else {
-	//If Bc charge is OK, write out the QQ indices for the two opposite-sign pairs 
-	if (fabs(charge) == 1){
-
-	  int mu_loneCharge = mu1_idx; int mu_SameCharge1 = mu2_idx; int mu_SameCharge2 = mu3_idx;
-	  //Look for the muon that has a different charge than the two others
-	  if (Reco_mu_charge[mu1_idx]==Reco_mu_charge[mu2_idx]){
-	    mu_loneCharge = mu3_idx; mu_SameCharge2 = mu1_idx;
-	  }
-	  else if (Reco_mu_charge[mu1_idx]==Reco_mu_charge[mu3_idx]){
-	    mu_loneCharge = mu2_idx; mu_SameCharge1 = mu1_idx;
-	  }
-
-	  //Look for the Jpsi indices corresponding to two muons
-	  Reco_3mu_QQ1_idx[Reco_3mu_size] = IndexOfThisJpsi(mu_loneCharge,mu_SameCharge1);
-	  Reco_3mu_QQ2_idx[Reco_3mu_size] = IndexOfThisJpsi(mu_loneCharge,mu_SameCharge2);
-	  if(Reco_3mu_QQ1_idx[Reco_3mu_size]==-1 && Reco_3mu_QQ2_idx[Reco_3mu_size]==-1) return; //need one OS pair to pass the dimuon selection
-	  Reco_3mu_QQss_idx[Reco_3mu_size] = IndexOfThisJpsi(mu_SameCharge1,mu_SameCharge2);
-
-	  //Split the muon types according to one valid QQ hypothesis among the two OS pairs
-	  if(Reco_3mu_QQ1_idx[Reco_3mu_size]>-1){ //The second os pair could also pass
-	    Reco_3mu_mumi_idx[Reco_3mu_size] = (Reco_mu_charge[mu_loneCharge]==-1)?mu_loneCharge:mu_SameCharge1;
-	    Reco_3mu_mupl_idx[Reco_3mu_size] = (Reco_mu_charge[mu_SameCharge1]==1)?mu_SameCharge1:mu_loneCharge;
-	    Reco_3mu_muW_idx[Reco_3mu_size] = mu_SameCharge2;
-	  }else{//Only the second pair is valid
-	    Reco_3mu_mumi_idx[Reco_3mu_size] = (Reco_mu_charge[mu_loneCharge]==-1)?mu_loneCharge:mu_SameCharge2;
-	    Reco_3mu_mupl_idx[Reco_3mu_size] = (Reco_mu_charge[mu_SameCharge2]==1)?mu_SameCharge2:mu_loneCharge;
-	    Reco_3mu_muW_idx[Reco_3mu_size] = mu_SameCharge1;
-	  }
-	  //pointing to the muW corresponding to the 2nd OS pair
-	  Reco_3mu_muW2_idx[Reco_3mu_size] = mu_SameCharge1;
-	
-	}//end Jpsi attribution for good Bc charge 
-
-	//If charge of the Bc is wrong, simpler procedure : random attribution of the three same-sign pairs
-	else {    
-	  Reco_3mu_QQ1_idx[Reco_3mu_size] = IndexOfThisJpsi(mu1_idx,mu2_idx);
-	  Reco_3mu_QQ2_idx[Reco_3mu_size] = IndexOfThisJpsi(mu1_idx,mu3_idx);
-	  Reco_3mu_QQss_idx[Reco_3mu_size] = IndexOfThisJpsi(mu2_idx,mu3_idx);
-
-	  Reco_3mu_mumi_idx[Reco_3mu_size] = mu1_idx;      //Which muon is mumi or mupl is random
-	  Reco_3mu_mupl_idx[Reco_3mu_size] = mu2_idx;
-	  Reco_3mu_muW_idx[Reco_3mu_size] = mu3_idx;
-	  Reco_3mu_muW2_idx[Reco_3mu_size] = mu2_idx;
-	}
-      } // end "no Jpsi flipping"
-
-      //*********
-      //Fill all remaining Bc variables 
-      Reco_3mu_charge[Reco_3mu_size] = charge;
-      TLorentzVector vBc = lorentzMomentum(aBcCand->p4());
-      new((*Reco_3mu_4mom)[Reco_3mu_size])TLorentzVector(vBc);
-      Reco_3mu_4mom_pt.push_back(vBc.Pt());
-      Reco_3mu_4mom_eta.push_back(vBc.Eta());
-      Reco_3mu_4mom_phi.push_back(vBc.Phi());
-      Reco_3mu_4mom_m.push_back(vBc.M());
-      // if(_flipJpsiDirection>0){
-      // 	cout<<"Bc mass, old vs new = "<< (vMuon1+vMuon2+vMuon3).M()<<" "<<vBc.M()<<endl;
-      // }
-
-      if (!(_isHI) && _muonLessPrimaryVertex && aBcCand->hasUserData("muonlessPV")) {
-	RefVtx = (*aBcCand->userData<reco::Vertex>("muonlessPV")).position();
-	RefVtx_xError = (*aBcCand->userData<reco::Vertex>("muonlessPV")).xError();
-	RefVtx_yError = (*aBcCand->userData<reco::Vertex>("muonlessPV")).yError();
-	RefVtx_zError = (*aBcCand->userData<reco::Vertex>("muonlessPV")).zError();
-      }
-      else if (aBcCand->hasUserData("PVwithmuons")) {
-	RefVtx = (*aBcCand->userData<reco::Vertex>("PVwithmuons")).position();
-	RefVtx_xError = (*aBcCand->userData<reco::Vertex>("PVwithmuons")).xError();
-	RefVtx_yError = (*aBcCand->userData<reco::Vertex>("PVwithmuons")).yError();
-	RefVtx_zError = (*aBcCand->userData<reco::Vertex>("PVwithmuons")).zError();
-      }
-      else {
-	cout << "HiOniaAnalyzer::fillTreeBc: no PVfor muon pair stored" << endl;
-	return;
-      }
-
-      new((*Reco_3mu_vtx)[Reco_3mu_size])TVector3(RefVtx.X(),RefVtx.Y(),RefVtx.Z());
-
-      
-      reco::Track iTrack_mupl, iTrack_mumi, iTrack_muW, mu1Trk, mu2Trk;
-      if(_flipJpsiDirection>0 && aBcCand->hasUserData("muon1Track") && aBcCand->hasUserData("muon2Track")){
-      	mu1Trk = *(aBcCand->userData<reco::Track>("muon1Track"));
-      	mu2Trk = *(aBcCand->userData<reco::Track>("muon2Track"));
-      }
-
-      //Long here but nothing much is happening -- just have to find back the mumi-pl-W indices
-      if ((!_theMinimumFlag && _muonLessPrimaryVertex) || (_flipJpsiDirection>0) ) {
-
-	if(_flipJpsiDirection>0){
-	  iTrack_mupl = (muon1->charge()>0)?mu1Trk:mu2Trk;
-	  iTrack_mumi = (muon1->charge()>0)?mu2Trk:mu1Trk;
-	  iTrack_muW = *(muon3->innerTrack());
-	}
-
-	else{ //find indices of mumi-pl-W
-	  if(Reco_3mu_muW_idx[Reco_3mu_size]==mu1_idx){
-	    iTrack_muW = *(muon1->innerTrack());
-	    if(Reco_3mu_mumi_idx[Reco_3mu_size]==mu2_idx){
-	      iTrack_mumi = *(muon2->innerTrack());
-	      iTrack_mupl = *(muon3->innerTrack());}
-	    else{
-	      iTrack_mumi = *(muon3->innerTrack());
-	      iTrack_mupl = *(muon2->innerTrack());}	      
-	  }
-	  else if(Reco_3mu_muW_idx[Reco_3mu_size]==mu2_idx){
-	    iTrack_muW = *(muon2->innerTrack());
-	    if(Reco_3mu_mumi_idx[Reco_3mu_size]==mu1_idx){
-	      iTrack_mumi = *(muon1->innerTrack());
-	      iTrack_mupl = *(muon3->innerTrack());}
-	    else{
-	      iTrack_mumi = *(muon3->innerTrack());
-	      iTrack_mupl = *(muon1->innerTrack());}	      
-	  }
-	  else {
-	    iTrack_muW = *(muon3->innerTrack());
-	    if(Reco_3mu_mumi_idx[Reco_3mu_size]==mu2_idx){
-	      iTrack_mumi = *(muon2->innerTrack());
-	      iTrack_mupl = *(muon1->innerTrack());}
-	    else{
-	      iTrack_mumi = *(muon1->innerTrack());
-	      iTrack_mupl = *(muon2->innerTrack());}	      
-	  }
-	}
-
-	Reco_3mu_muW_dxy[Reco_3mu_size] = iTrack_muW.dxy(RefVtx);
-	Reco_3mu_muW_dz[Reco_3mu_size] = iTrack_muW.dz(RefVtx);
-	Reco_3mu_mumi_dxy[Reco_3mu_size] = iTrack_mumi.dxy(RefVtx);
-	Reco_3mu_mumi_dz[Reco_3mu_size] = iTrack_mumi.dz(RefVtx);
-	Reco_3mu_mupl_dxy[Reco_3mu_size] = iTrack_mupl.dxy(RefVtx);
-	Reco_3mu_mupl_dz[Reco_3mu_size] = iTrack_mupl.dz(RefVtx);
-      }
-
-      //*********
-      //Lifetime related variables
-
-      if (_useBS) {
-	if (aBcCand->hasUserFloat("ppdlBS")) {
-	  Reco_3mu_ctau[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlBS");
-	} else {  
-	  Reco_3mu_ctau[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlBS was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlErrBS")) {
-	  Reco_3mu_ctauErr[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlErrBS");
-	} else {
-	  Reco_3mu_ctauErr[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlErrBS was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlBS3D")) {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlBS3D");
-	} else {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlBS3D was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlErrBS3D")) {
-	  Reco_3mu_ctauErr3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlErrBS3D");
-	} else {
-	  Reco_3mu_ctauErr3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlErrBS3D was not found" << std::endl;
-	}
-      }
-      else {
-	if (aBcCand->hasUserFloat("ppdlPV")) {
-	  Reco_3mu_ctau[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlPV");
-	} else {
-	  Reco_3mu_ctau[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlPV was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlErrPV")) {
-	  Reco_3mu_ctauErr[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlErrPV");
-	} else {
-	  Reco_3mu_ctauErr[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlErrPV was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlPV3D")) {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlPV3D");
-	} else {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlPV3D was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlErrPV3D")) {
-	  Reco_3mu_ctauErr3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlErrPV3D");
-	} else {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlErrPV3D was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("cosAlpha")) {
-	  Reco_3mu_cosAlpha[Reco_3mu_size] = aBcCand->userFloat("cosAlpha");
-	} else {
-	  Reco_3mu_cosAlpha[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float cosAlpha was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("cosAlpha3D")) {
-	  Reco_3mu_cosAlpha3D[Reco_3mu_size] = aBcCand->userFloat("cosAlpha3D");
-	} else {
-	  Reco_3mu_cosAlpha3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float cosAlpha3D was not found" << std::endl;
-	}
-      }
-
-      if (aBcCand->hasUserFloat("vProb")) {
-	Reco_3mu_VtxProb[Reco_3mu_size] = aBcCand->userFloat("vProb");
-      } else {
-	Reco_3mu_VtxProb[Reco_3mu_size] = -1;
-	std::cout << "Warning: User Float vProb was not found" << std::endl;
-      }
-      if (aBcCand->hasUserFloat("MassErr")) {
-	Reco_3mu_MassErr[Reco_3mu_size] = aBcCand->userFloat("MassErr");
-      } else {
-	Reco_3mu_MassErr[Reco_3mu_size] = -10;
-	std::cout << "Warning: User Float MassErr was not found" << std::endl;
-      }
-
-      //Correct the Bc mass for the momentum of the neutrino, transverse to the Bc flight direction
-      float Mtrimu = vBc.M();
-      float Ptrimu = vBc.P();
-      float sinalpha = sin(acos(Reco_3mu_cosAlpha3D[Reco_3mu_size]));
-      float PperpTrimu = sinalpha * Ptrimu;
-      Reco_3mu_CorrM[Reco_3mu_size] = sqrt(Mtrimu*Mtrimu + PperpTrimu*PperpTrimu) + PperpTrimu;
-
-      if(_useSVfinder && SVs.isValid() && SVs->size()>0){
-	Reco_3mu_NbMuInSameSV[Reco_3mu_size] = MuInSV(vMuon1,vMuon2,vMuon3);
-      }
-
-      Reco_3mu_size++;
-    }
-  }
-
-  return;
-}
-
-void
-HiOniaAnalyzer::fillTreeDimuTrk(int count) {
-
-  if (Reco_3mu_size >= Max_Bc_size) {
-    std::cout << "Too many dimuon+track candidates: " << Reco_3mu_size << std::endl;
-    std::cout << "Maximum allowed: " << Max_Bc_size << std::endl;
-    return;
-  }
-
-  const pat::CompositeCandidate* aBcCand = _thePassedBcCands.at(count);
-  
-  if (aBcCand==NULL){
-    std::cout<<"ERROR: 'aBcCand' pointer in fillTreeDimuTrk is NULL ! Return now"<<std::endl; return;
-  } else {
-    const pat::Muon* muon1 = dynamic_cast<const pat::Muon*>(aBcCand->daughter("muon1"));
-    const pat::Muon* muon2 = dynamic_cast<const pat::Muon*>(aBcCand->daughter("muon2"));
-    const reco::RecoChargedCandidate* trk3 = dynamic_cast<const reco::RecoChargedCandidate*>(aBcCand->daughter("track"));
-	
-    if (muon1==NULL || muon2==NULL || trk3==NULL){
-      std::cout<<"ERROR: 'muon1' or 'muon2' or 'trk3' pointer in fillTreeDimuTrk is NULL ! Return now"<<std::endl; return;
-    } else {
-
-      Reco_3mu_charge[Reco_3mu_size] = muon1->charge() + muon2->charge()+ trk3->charge(); 
-      
-      TLorentzVector vMuon1 = lorentzMomentum(muon1->p4());
-      TLorentzVector vMuon2 = lorentzMomentum(muon2->p4());
-      TLorentzVector vTrk3 = lorentzMomentum(trk3->p4());
-
-      int mu1_idx = IndexOfThisMuon(&vMuon1);
-      int mu2_idx = IndexOfThisMuon(&vMuon2);
-      int trk3_idx = IndexOfThisTrack(&vTrk3);
-
-      //The dimuon has to pass the Jpsi kinematic cuts
-      Reco_3mu_QQ1_idx[Reco_3mu_size] = IndexOfThisJpsi(mu1_idx,mu2_idx); 
-      if (Reco_3mu_QQ1_idx[Reco_3mu_size]==-1) {return;}
-
-      TLorentzVector vBc = lorentzMomentum(aBcCand->p4());
-      new((*Reco_3mu_4mom)[Reco_3mu_size])TLorentzVector(vBc);
-      Reco_3mu_4mom_pt.push_back(vBc.Pt());
-      Reco_3mu_4mom_eta.push_back(vBc.Eta());
-      Reco_3mu_4mom_phi.push_back(vBc.Phi());
-      Reco_3mu_4mom_m.push_back(vBc.M());
-
-      Reco_3mu_mumi_idx[Reco_3mu_size] = (Reco_mu_charge[mu1_idx]==-1)?mu1_idx:mu2_idx;
-      Reco_3mu_mupl_idx[Reco_3mu_size] = (Reco_mu_charge[mu1_idx]==-1)?mu2_idx:mu1_idx;
-      Reco_3mu_muW_idx[Reco_3mu_size] = trk3_idx; //let's keep the same variables names as for the trimuon (for internal use...)
-      
-      //*********
-      //Fill all remaining Bc variables 
-      if (!(_isHI) && _muonLessPrimaryVertex && aBcCand->hasUserData("muonlessPV")) {
-	RefVtx = (*aBcCand->userData<reco::Vertex>("muonlessPV")).position();
-	RefVtx_xError = (*aBcCand->userData<reco::Vertex>("muonlessPV")).xError();
-	RefVtx_yError = (*aBcCand->userData<reco::Vertex>("muonlessPV")).yError();
-	RefVtx_zError = (*aBcCand->userData<reco::Vertex>("muonlessPV")).zError();
-      }
-      else if (aBcCand->hasUserData("PVwithmuons")) {
-	RefVtx = (*aBcCand->userData<reco::Vertex>("PVwithmuons")).position();
-	RefVtx_xError = (*aBcCand->userData<reco::Vertex>("PVwithmuons")).xError();
-	RefVtx_yError = (*aBcCand->userData<reco::Vertex>("PVwithmuons")).yError();
-	RefVtx_zError = (*aBcCand->userData<reco::Vertex>("PVwithmuons")).zError();
-      }
-      else {
-	cout << "HiOniaAnalyzer::fillTreeDimuTrk: no PVfor muon pair stored" << endl;
-	return;
-      }
-
-      new((*Reco_3mu_vtx)[Reco_3mu_size])TVector3(RefVtx.X(),RefVtx.Y(),RefVtx.Z());
-
-      Reco_3mu_muW_dxy[Reco_3mu_size] = trk3->track()->dxy(RefVtx);
-      Reco_3mu_muW_dz[Reco_3mu_size] = trk3->track()->dz(RefVtx);
-
-      //*********
-      //Lifetime related variables
-
-      if (_useBS) {
-	if (aBcCand->hasUserFloat("ppdlBS")) {
-	  Reco_3mu_ctau[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlBS");
-	} else {  
-	  Reco_3mu_ctau[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlBS was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlErrBS")) {
-	  Reco_3mu_ctauErr[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlErrBS");
-	} else {
-	  Reco_3mu_ctauErr[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlErrBS was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlBS3D")) {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlBS3D");
-	} else {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlBS3D was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlErrBS3D")) {
-	  Reco_3mu_ctauErr3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlErrBS3D");
-	} else {
-	  Reco_3mu_ctauErr3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlErrBS3D was not found" << std::endl;
-	}
-      }
-      else {
-	if (aBcCand->hasUserFloat("ppdlPV")) {
-	  Reco_3mu_ctau[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlPV");
-	} else {
-	  Reco_3mu_ctau[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlPV was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlErrPV")) {
-	  Reco_3mu_ctauErr[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlErrPV");
-	} else {
-	  Reco_3mu_ctauErr[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlErrPV was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlPV3D")) {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlPV3D");
-	} else {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlPV3D was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("ppdlErrPV3D")) {
-	  Reco_3mu_ctauErr3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("ppdlErrPV3D");
-	} else {
-	  Reco_3mu_ctau3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float ppdlErrPV3D was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("cosAlpha")) {
-	  Reco_3mu_cosAlpha[Reco_3mu_size] = aBcCand->userFloat("cosAlpha");
-	} else {
-	  Reco_3mu_cosAlpha[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float cosAlpha was not found" << std::endl;
-	}
-	if (aBcCand->hasUserFloat("cosAlpha3D")) {
-	  Reco_3mu_cosAlpha3D[Reco_3mu_size] = aBcCand->userFloat("cosAlpha3D");
-	} else {
-	  Reco_3mu_cosAlpha3D[Reco_3mu_size] = -10;
-	  std::cout << "Warning: User Float cosAlpha3D was not found" << std::endl;
-	}
-
-	if(_doDimuTrk){
-	  if (aBcCand->hasUserFloat("KCppdlPV")) {
-	    Reco_3mu_KCctau[Reco_3mu_size] = 10.0*aBcCand->userFloat("KCppdlPV");
-	  } else {
-	    Reco_3mu_KCctau[Reco_3mu_size] = -10;
-	    std::cout << "Warning: User Float KCppdlPV was not found" << std::endl;
-	  }
-	  if (aBcCand->hasUserFloat("KCppdlErrPV")) {
-	    Reco_3mu_KCctauErr[Reco_3mu_size] = 10.0*aBcCand->userFloat("KCppdlErrPV");
-	  } else {
-	    Reco_3mu_KCctauErr[Reco_3mu_size] = -10;
-	    std::cout << "Warning: User Float KCppdlErrPV was not found" << std::endl;
-	  }
-	  if (aBcCand->hasUserFloat("KCppdlPV3D")) {
-	    Reco_3mu_KCctau3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("KCppdlPV3D");
-	  } else {
-	    Reco_3mu_KCctau3D[Reco_3mu_size] = -10;
-	    std::cout << "Warning: User Float KCppdlPV3D was not found" << std::endl;
-	  }
-	  if (aBcCand->hasUserFloat("KCppdlErrPV3D")) {
-	    Reco_3mu_KCctauErr3D[Reco_3mu_size] = 10.0*aBcCand->userFloat("KCppdlErrPV3D");
-	  } else {
-	    Reco_3mu_KCctau3D[Reco_3mu_size] = -10;
-	    std::cout << "Warning: User Float KCppdlErrPV3D was not found" << std::endl;
-	  }
-	  if (aBcCand->hasUserFloat("KCcosAlpha")) {
-	    Reco_3mu_KCcosAlpha[Reco_3mu_size] = aBcCand->userFloat("KCcosAlpha");
-	  } else {
-	    Reco_3mu_KCcosAlpha[Reco_3mu_size] = -10;
-	    std::cout << "Warning: User Float KCcosAlpha was not found" << std::endl;
-	  }
-	  if (aBcCand->hasUserFloat("KCcosAlpha3D")) {
-	    Reco_3mu_KCcosAlpha3D[Reco_3mu_size] = aBcCand->userFloat("KCcosAlpha3D");
-	  } else {
-	    Reco_3mu_KCcosAlpha3D[Reco_3mu_size] = -10;
-	    std::cout << "Warning: User Float KCcosAlpha3D was not found" << std::endl;
-	  }
-	}
-
-      }
-
-      if (aBcCand->hasUserFloat("vProb")) {
-	Reco_3mu_VtxProb[Reco_3mu_size] = aBcCand->userFloat("vProb");
-      } else {
-	Reco_3mu_VtxProb[Reco_3mu_size] = -1;
-	std::cout << "Warning: User Float vProb was not found" << std::endl;
-      }
-      if (aBcCand->hasUserFloat("KinConstrainedVtxProb")) {
-	Reco_3mu_KCVtxProb[Reco_3mu_size] = aBcCand->userFloat("KinConstrainedVtxProb");
-      } else {
-	Reco_3mu_KCVtxProb[Reco_3mu_size] = -10;
-	std::cout << "Warning: User Float KinConstrainedVtxProb was not found" << std::endl;
-      }
-      if (aBcCand->hasUserFloat("MassErr")) {
-	Reco_3mu_MassErr[Reco_3mu_size] = aBcCand->userFloat("MassErr");
-      } else {
-	Reco_3mu_MassErr[Reco_3mu_size] = -10;
-	std::cout << "Warning: User Float MassErr was not found" << std::endl;
-      }
-
-      //Correct the Bc mass for the momentum of the neutrino, transverse to the Bc flight direction
-      float Mtrimu = vBc.M();
-      float Ptrimu = vBc.P();
-      float sinalpha = sin(acos(Reco_3mu_cosAlpha3D[Reco_3mu_size]));
-      float PperpTrimu = sinalpha * Ptrimu;
-      Reco_3mu_CorrM[Reco_3mu_size] = sqrt(Mtrimu*Mtrimu + PperpTrimu*PperpTrimu) + PperpTrimu;
-
-      if(_useSVfinder && SVs.isValid() && SVs->size()>0){
-	Reco_3mu_NbMuInSameSV[Reco_3mu_size] = MuInSV(vMuon1,vMuon2,vTrk3);
-      }
-      
-      Reco_3mu_size++;
-    }
-  }
-
-  return;
-}
-
-void
-HiOniaAnalyzer::fillRecoJpsi(int count, std::string trigName, std::string centName) {
+void HiOniaAnalyzer::fillRecoJpsi(int count, std::string trigName, std::string centName){
   pat::CompositeCandidate* aJpsiCand = _thePassedCands.at(count)->clone();
 
   if(aJpsiCand==NULL){
@@ -2043,16 +916,13 @@ HiOniaAnalyzer::fillRecoJpsi(int count, std::string trigName, std::string centNa
 
   delete aJpsiCand;
   return;
-}
+};
 
-void
-HiOniaAnalyzer::fillHistosAndDS(unsigned int theCat, const pat::CompositeCandidate* aJpsiCand) {
-
+void HiOniaAnalyzer::fillHistosAndDS(unsigned int theCat, const pat::CompositeCandidate* aJpsiCand){
   return;
-}
+};
 
-void
-HiOniaAnalyzer::checkTriggers(const pat::CompositeCandidate* aJpsiCand) {
+void HiOniaAnalyzer::checkTriggers(const pat::CompositeCandidate* aJpsiCand){
 
   if(aJpsiCand==NULL){
     std::cout<<"ERROR: 'aJpsiCand' pointer in checkTriggers is NULL ! Return now"<<std::endl; return;
@@ -2077,9 +947,6 @@ HiOniaAnalyzer::checkTriggers(const pat::CompositeCandidate* aJpsiCand) {
 	bool pass1 = mu1HLTMatchesFilter.size() > 0;
 	bool pass2 = mu2HLTMatchesFilter.size() > 0;
     
-	//    pass1 = mu1HLTMatchesPath.size() > 0;
-	//    pass2 = mu2HLTMatchesPath.size() > 0;
-    
 	if (iTr > NTRIGGERS_DBL) {  // single triggers here
 	  isTriggerMatched[iTr] = pass1 || pass2;
 	} else {        // double triggers here
@@ -2096,357 +963,10 @@ HiOniaAnalyzer::checkTriggers(const pat::CompositeCandidate* aJpsiCand) {
       }
     }
   }
-
   return;
-}
+};
 
-void
-HiOniaAnalyzer::makeCuts(bool keepSameSign) {
-  math::XYZPoint RefVtx_tmp = RefVtx;
-
-  if (collJpsi.isValid()) {
-
-    for(std::vector<pat::CompositeCandidate>::const_iterator it=collJpsi->begin();
-        it!=collJpsi->end(); ++it) {
-      
-      const pat::CompositeCandidate* cand = &(*it);     
-
-      if(cand==NULL){
-	std::cout<<"ERROR: 'cand' pointer in makeCuts is NULL ! Return now"<<std::endl; return;
-      } else{
-
-        const pat::Muon* muon1 = dynamic_cast<const pat::Muon*>(cand->daughter("muon1"));
-        const pat::Muon* muon2 = dynamic_cast<const pat::Muon*>(cand->daughter("muon2"));
-
-        if (muon1==NULL || muon2==NULL){
-	  std::cout<<"ERROR: 'muon1' or 'muon2' pointer in makeCuts is NULL ! Return now"<<std::endl; return;
-        } else {
-
-          if(!keepSameSign && (muon1->charge() + muon2->charge() != 0)) continue;
-
-	  if (!(_isHI) && _muonLessPrimaryVertex && cand->hasUserData("muonlessPV"))
-	    RefVtx = (*cand->userData<reco::Vertex>("muonlessPV")).position();
-	  else if (!_muonLessPrimaryVertex && cand->hasUserData("PVwithmuons"))
-	    RefVtx = (*cand->userData<reco::Vertex>("PVwithmuons")).position();
-	  else {
-	    std::cout << "HiOniaAnalyzer::makeCuts: no PV for muon pair stored ! Go to next candidate." << std::endl;
-	    continue;
-	  }
-
-	  if (fabs(RefVtx.Z()) > _iConfig.getParameter< double > ("maxAbsZ")) continue;
-      
-	  if (fabs(muon1->eta()) >= etaMax ||
-	      fabs(muon2->eta()) >= etaMax) continue;
-
-	  //Pass muon selection? 
-	  bool muonSelFound = false;
-	  if ( _muonSel==(std::string)("Glb") ) {
-	    if (checkCuts(cand,muon1,muon2,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selGlobalMuon)){
-	      _thePassedCats.push_back(Glb_Glb);  _thePassedCands.push_back(cand);
-	      if(!_fillSingleMuons){
-		EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());}
-	    }
-	    muonSelFound = true;
-	  }
-	  if ( _muonSel==(std::string)("TwoGlbAmongThree") ) {
-	    if (checkCuts(cand,muon1,muon2,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selTrackerMuon)){
-	      _thePassedCats.push_back(TwoGlbAmongThree);  _thePassedCands.push_back(cand);
-	      if(!_fillSingleMuons){
-		EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());}
-	    }
-	    muonSelFound = true;
-	  }
-	  if ( _muonSel==(std::string)("GlbTrk") ) {
-	    if (checkCuts(cand,muon1,muon2,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selGlobalMuon)){
-	      _thePassedCats.push_back(GlbTrk_GlbTrk);  _thePassedCands.push_back(cand);
-	      if(!_fillSingleMuons){
-		EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());}
-	    }
-	    muonSelFound = true;
-	  }
-	  if ( _muonSel==(std::string)("Trk") ) {
-	    if (checkCuts(cand,muon1,muon2,&HiOniaAnalyzer::selTrackerMuon,&HiOniaAnalyzer::selTrackerMuon)){
-	      _thePassedCats.push_back(Trk_Trk);  _thePassedCands.push_back(cand);
-	      if(!_fillSingleMuons){
-		EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());}
-	    }
-	    muonSelFound = true;
-	  }
-          if ( _muonSel==(std::string)("GlbOrTrk") ){
-	    if (checkCuts(cand,muon1,muon2,&HiOniaAnalyzer::selGlobalOrTrackerMuon,&HiOniaAnalyzer::selGlobalOrTrackerMuon)){
-	      _thePassedCats.push_back(GlbOrTrk_GlbOrTrk);  _thePassedCands.push_back(cand);
-	      if(!_fillSingleMuons){
-		EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());}
-	    }
-	    muonSelFound = true;
-          }
-          if ( _muonSel==(std::string)("All") ){
-	    if (checkCuts(cand,muon1,muon2,&HiOniaAnalyzer::selAllMuon,&HiOniaAnalyzer::selAllMuon)){
-	      _thePassedCats.push_back(All_All);  _thePassedCands.push_back(cand);
-	      if(!_fillSingleMuons){
-		EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());}
-	    }
-	    muonSelFound = true;
-          }
-	  if (!muonSelFound) {
-	    std::cout << "[HiOniaAnalyzer::makeCuts] --- The muon selection: " << _muonSel << " is invalid. The supported options are: All, Glb, GlbTrk, GlbOrTrk, Trk, and TwoGlbAmongThree" << std::endl;
-	  }
-	}
-      }
-    }
-  }  
-  RefVtx = RefVtx_tmp;
-  return;
-}
-
-void
-HiOniaAnalyzer::makeBcCuts(bool keepWrongSign) {
-  math::XYZPoint RefVtx_tmp = RefVtx;
-
-  if (collTrimuon.isValid()) {
-
-    for(std::vector<pat::CompositeCandidate>::const_iterator it=collTrimuon->begin();
-        it!=collTrimuon->end(); ++it) {
-
-      const pat::CompositeCandidate* cand = &(*it);     
-
-      if(cand==NULL){
-	std::cout<<"ERROR: 'cand' pointer in makeBcCuts is NULL ! Return now"<<std::endl; return;
-      } else{
-    
-	const pat::Muon* muon1 = dynamic_cast<const pat::Muon*>(cand->daughter("muon1"));
-	const pat::Muon* muon2 = dynamic_cast<const pat::Muon*>(cand->daughter("muon2"));
-	const pat::Muon* muon3 = dynamic_cast<const pat::Muon*>(cand->daughter("muon3"));
-
-	if (muon1==NULL || muon2==NULL || muon3==NULL){
-	  std::cout<<"ERROR: 'muon1' or 'muon2' or 'muon3' pointer in makeBcCuts is NULL ! Return now"<<std::endl; return;
-	} else {  
-	
-	  if(!keepWrongSign && (muon1->charge() + muon2->charge() + muon3->charge() != 1) 
-	     && (muon1->charge() + muon2->charge() + muon3->charge() != -1)) continue;
-
-	  if (!(_isHI) && _muonLessPrimaryVertex && cand->hasUserData("muonlessPV"))
-	    RefVtx = (*cand->userData<reco::Vertex>("muonlessPV")).position();
-	  else if (!_muonLessPrimaryVertex && cand->hasUserData("PVwithmuons"))
-	    RefVtx = (*cand->userData<reco::Vertex>("PVwithmuons")).position();
-	  else {
-	    std::cout << "HiOniaAnalyzer::makeCuts: no PV for muon pair stored ! Go to next candidate." << std::endl;
-	    continue;
-	  }
-      
-	  if (fabs(RefVtx.Z()) > _iConfig.getParameter< double > ("maxAbsZ")) continue;
-      
-	  if (fabs(muon1->eta()) >= etaMax || fabs(muon2->eta()) >= etaMax || fabs(muon3->eta()) >= etaMax ) continue;
-      
-	  //Pass muon selection?
-	  if (    ( _muonSel==(std::string)("GlbOrTrk") ) &&
-		  checkBcCuts(cand,muon1,muon2,muon3,&HiOniaAnalyzer::selGlobalOrTrackerMuon,&HiOniaAnalyzer::selGlobalOrTrackerMuon,&HiOniaAnalyzer::selGlobalOrTrackerMuon)
-		  ){
-	    _thePassedBcCats.push_back(Glb_Glb);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta()); EtaOfWantedMuons.push_back(muon3->eta());}
-	  }
-	  else if(( _muonSel==(std::string)("TwoGlbAmongThree") ) && //argument functions 2 and 3 have to be the same for good symmetrization
-		  checkBcCuts(cand,muon1,muon2,muon3,&HiOniaAnalyzer::selTrackerMuon,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selGlobalMuon)
-		  ){
-	    _thePassedBcCats.push_back(TwoGlbAmongThree);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta()); EtaOfWantedMuons.push_back(muon3->eta());}
-	  }
-	  else if(( _muonSel==(std::string)("Glb") ) &&
-		  checkBcCuts(cand,muon1,muon2,muon3,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selGlobalMuon)
-		  ){
-	    _thePassedBcCats.push_back(Glb_Glb);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta()); EtaOfWantedMuons.push_back(muon3->eta());}
-	  }
-	  else if(( _muonSel==(std::string)("GlbTrk") ) &&
-		  checkBcCuts(cand,muon1,muon2,muon3,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selGlobalMuon)
-		  ){
-	    _thePassedBcCats.push_back(GlbTrk_GlbTrk);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta()); EtaOfWantedMuons.push_back(muon3->eta());}
-	  }
-	  else if(( _muonSel==(std::string)("Trk") ) &&
-		  checkBcCuts(cand,muon1,muon2,muon3,&HiOniaAnalyzer::selTrackerMuon,&HiOniaAnalyzer::selTrackerMuon,&HiOniaAnalyzer::selTrackerMuon)
-		  ){
-	    _thePassedBcCats.push_back(Trk_Trk);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta()); EtaOfWantedMuons.push_back(muon3->eta());}
-	  }
-	  else if(( _muonSel==(std::string)("All") ) &&
-		  checkBcCuts(cand,muon1,muon2,muon3,&HiOniaAnalyzer::selAllMuon,&HiOniaAnalyzer::selAllMuon,&HiOniaAnalyzer::selAllMuon)
-		  ){
-	    _thePassedBcCats.push_back(All_All);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta()); EtaOfWantedMuons.push_back(muon3->eta());}
-	  }
-	  else {
-	    //std::cout << "[HiOniaAnalyzer::makeCuts] trimuon --- muon did not pass selection: " << _muonSel << std::endl;
-	  }
-	}
-      }
-    }
-  }
-  
-  RefVtx = RefVtx_tmp;
-  return;
-}
-
-void
-HiOniaAnalyzer::makeDimutrkCuts(bool keepWrongSign) {
-  math::XYZPoint RefVtx_tmp = RefVtx;
-
-  if (collDimutrk.isValid()) {
-    
-    for(std::vector<pat::CompositeCandidate>::const_iterator it=collDimutrk->begin();
-        it!=collDimutrk->end(); ++it) {
-
-      const pat::CompositeCandidate* cand = &(*it);     
-
-      if(cand==NULL){
-	std::cout<<"ERROR: 'cand' pointer in makeDimutrkCuts is NULL ! Return now"<<std::endl; return;
-      } else{
-    
-	const pat::Muon* muon1 = dynamic_cast<const pat::Muon*>(cand->daughter("muon1"));
-	const pat::Muon* muon2 = dynamic_cast<const pat::Muon*>(cand->daughter("muon2"));
-	const reco::RecoChargedCandidate* trk = dynamic_cast<const reco::RecoChargedCandidate*>(cand->daughter("track"));
-
-	if (muon1==NULL || muon2==NULL || trk==NULL){
-	  std::cout<<"ERROR: 'muon1' or 'muon2' or 'trk' pointer in makeDimutrkCuts is NULL ! Return now"<<std::endl; return;
-	} else {  
-	
-	  if(!keepWrongSign && (muon1->charge() + muon2->charge() + trk->charge() != 1) 
-	     && (muon1->charge() + muon2->charge() + trk->charge() != -1)) continue;
-
-	  if (!(_isHI) && _muonLessPrimaryVertex && cand->hasUserData("muonlessPV"))
-	    RefVtx = (*cand->userData<reco::Vertex>("muonlessPV")).position();
-	  else if (!_muonLessPrimaryVertex && cand->hasUserData("PVwithmuons"))
-	    RefVtx = (*cand->userData<reco::Vertex>("PVwithmuons")).position();
-	  else {
-	    std::cout << "HiOniaAnalyzer::makeCuts: no PV for muon pair stored ! Go to next candidate." << std::endl;
-	    continue;
-	  }
-      
-	  if (fabs(RefVtx.Z()) > _iConfig.getParameter< double > ("maxAbsZ")) continue;
-      
-	  if (fabs(muon1->eta()) >= etaMax || fabs(muon2->eta()) >= etaMax || fabs(trk->eta()) >= etaMax ) continue;
-      
-	  //Pass muon selection?
-	  if (    ( _muonSel==(std::string)("GlbOrTrk") ) &&
-		  checkDimuTrkCuts(cand,muon1,muon2,trk,&HiOniaAnalyzer::selGlobalOrTrackerMuon,&HiOniaAnalyzer::selGlobalOrTrackerMuon,&HiOniaAnalyzer::selTrk)
-		  ){
-	    _thePassedBcCats.push_back(Glb_Glb);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());} 
-	    EtaOfWantedTracks.push_back(trk->eta());
-	  }
-	  else if(( _muonSel==(std::string)("TwoGlbAmongThree") ) && //argument functions 2 and 3 have to be the same for good symmetrization
-		  checkDimuTrkCuts(cand,muon1,muon2,trk,&HiOniaAnalyzer::selTrackerMuon,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selTrk)
-		  ){
-	    _thePassedBcCats.push_back(TwoGlbAmongThree);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());} 
-	    EtaOfWantedTracks.push_back(trk->eta());
-	  }
-	  else if(( _muonSel==(std::string)("Glb") ) &&
-		  checkDimuTrkCuts(cand,muon1,muon2,trk,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selTrk)
-		  ){
-	    _thePassedBcCats.push_back(Glb_Glb);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());} 
-	    EtaOfWantedTracks.push_back(trk->eta());
-	  }
-	  else if(( _muonSel==(std::string)("GlbTrk") ) &&
-		  checkDimuTrkCuts(cand,muon1,muon2,trk,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selGlobalMuon,&HiOniaAnalyzer::selTrk)
-		  ){
-	    _thePassedBcCats.push_back(GlbTrk_GlbTrk);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());}
-	    EtaOfWantedTracks.push_back(trk->eta());
-	  }
-	  else if(( _muonSel==(std::string)("Trk") ) &&
-		  checkDimuTrkCuts(cand,muon1,muon2,trk,&HiOniaAnalyzer::selTrackerMuon,&HiOniaAnalyzer::selTrackerMuon,&HiOniaAnalyzer::selTrk)
-		  ){
-	    _thePassedBcCats.push_back(Trk_Trk);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());} 
-	    EtaOfWantedTracks.push_back(trk->eta());
-	  }
-	  else if(( _muonSel==(std::string)("All") ) &&
-		  checkDimuTrkCuts(cand,muon1,muon2,trk,&HiOniaAnalyzer::selAllMuon,&HiOniaAnalyzer::selAllMuon,&HiOniaAnalyzer::selTrk)
-		  ){
-	    _thePassedBcCats.push_back(All_All);  _thePassedBcCands.push_back(cand);
-	    if(!_fillSingleMuons){
-	      EtaOfWantedMuons.push_back(muon1->eta()); EtaOfWantedMuons.push_back(muon2->eta());} 
-	    EtaOfWantedTracks.push_back(trk->eta());
-	  }
-	  else {
-	    //std::cout << "[HiOniaAnalyzer::makeCuts] trimuon --- muon did not pass selection: " << _muonSel << std::endl;
-	  }
-	}
-      }
-    }
-  }
-  
-  RefVtx = RefVtx_tmp;
-  return;
-}
-
-bool
-HiOniaAnalyzer::checkCuts(const pat::CompositeCandidate* cand, const pat::Muon* muon1,  const pat::Muon* muon2, bool(HiOniaAnalyzer::* callFunc1)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc2)(const pat::Muon*)) {
-  std::string lastFilter = _OneMatchedHLTMu>=0 ? filterNameMap.at(theTriggerNames[_OneMatchedHLTMu]) : "";
-  if ( (((this->*callFunc1)(muon1) && (this->*callFunc2)(muon2)) || ((this->*callFunc1)(muon2) && (this->*callFunc2)(muon1))) &&
-       (!_applycuts || true) && //Add hard-coded cuts here if desired
-       (    (_OneMatchedHLTMu==-1) || 
-	    (muon1->triggerObjectMatchesByFilter(lastFilter)).size()>0 || (muon2->triggerObjectMatchesByFilter(lastFilter)).size()>0
-	    ) ) 
-    return true;
-  else
-    return false;
-}
-
-bool
-HiOniaAnalyzer::checkBcCuts(const pat::CompositeCandidate* cand, const pat::Muon* muon1, const pat::Muon* muon2, const pat::Muon* muon3, bool(HiOniaAnalyzer::* callFunc1)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc2)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc3)(const pat::Muon*)) {
-  const auto& lastFilter = filterNameMap.at(theTriggerNames[(_OneMatchedHLTMu<0)?0:_OneMatchedHLTMu]);
-  const auto& mu1HLTMatchesFilter = muon1->triggerObjectMatchesByFilter( lastFilter );
-  const auto& mu2HLTMatchesFilter = muon2->triggerObjectMatchesByFilter( lastFilter );
-  const auto& mu3HLTMatchesFilter = muon3->triggerObjectMatchesByFilter( lastFilter );
-  
-  if ( ( ((this->*callFunc1)(muon1) && (this->*callFunc2)(muon2) && (this->*callFunc3)(muon3))
-         //symmetrize, assuming arguments functions 2 and 3 are THE SAME ! 
-         || ((this->*callFunc1)(muon2) && (this->*callFunc2)(muon1) && (this->*callFunc3)(muon3))
-         || ((this->*callFunc1)(muon3) && (this->*callFunc2)(muon1) && (this->*callFunc3)(muon2)) ) &&
-       (!_applycuts || true) && //Add hard-coded cuts here if desired 
-       (    (_OneMatchedHLTMu==-1) || 
-	    (mu1HLTMatchesFilter.size() > 0 && mu2HLTMatchesFilter.size() > 0) ||
-	    (mu1HLTMatchesFilter.size() > 0 && mu3HLTMatchesFilter.size() > 0) ||
-	    (mu2HLTMatchesFilter.size() > 0 && mu3HLTMatchesFilter.size() > 0)
-	    ) )
-    return true;
-  else
-    return false;
-}
-
-bool
-HiOniaAnalyzer::checkDimuTrkCuts(const pat::CompositeCandidate* cand, const pat::Muon* muon1, const pat::Muon* muon2, const reco::RecoChargedCandidate* trk, bool(HiOniaAnalyzer::* callFunc1)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc2)(const pat::Muon*), bool(HiOniaAnalyzer::* callFunc3)(const reco::TrackRef)) {
-  const auto& lastFilter = filterNameMap.at(theTriggerNames[(_OneMatchedHLTMu<0)?0:_OneMatchedHLTMu]);
-  const auto& mu1HLTMatchesFilter = muon1->triggerObjectMatchesByFilter( lastFilter );
-  const auto& mu2HLTMatchesFilter = muon2->triggerObjectMatchesByFilter( lastFilter );
-  
-  if ( ( ((this->*callFunc1)(muon1) && (this->*callFunc2)(muon2) && (this->*callFunc3)(trk->track()))
-         || ((this->*callFunc1)(muon2) && (this->*callFunc2)(muon1) && (this->*callFunc3)(trk->track())) ) &&
-       (!_applycuts || true) && //Add hard-coded cuts here if desired 
-       (    (_OneMatchedHLTMu==-1) || 
-	    (mu1HLTMatchesFilter.size() > 0 && mu2HLTMatchesFilter.size() > 0)
-	    ) )
-    return true;
-  else
-    return false;
-}
-
-pair< unsigned int, const pat::CompositeCandidate* > 
-HiOniaAnalyzer::theBestQQ() {
+pair< unsigned int, const pat::CompositeCandidate* >  HiOniaAnalyzer::theBestQQ() {
 
   unsigned int theBestCat = 99;
   const pat::CompositeCandidate* theBestCand = new pat::CompositeCandidate();
@@ -2460,169 +980,14 @@ HiOniaAnalyzer::theBestQQ() {
 
   pair< unsigned int, const pat::CompositeCandidate* > result = make_pair(theBestCat, theBestCand );
   return result;
-}
+};
 
-bool 
-HiOniaAnalyzer::isTrkInMuonAccept(TLorentzVector trk4mom, std::string muonType){
-  if (muonType == (std::string)("GLB")) {
-    return (fabs(trk4mom.Eta()) < 2.4 &&
-            ((fabs(trk4mom.Eta()) < 1.2 && trk4mom.Pt() >= 3.5) ||
-             (1.2 <= fabs(trk4mom.Eta()) && fabs(trk4mom.Eta()) < 2.1 && trk4mom.Pt() >= 5.47-1.89*fabs(trk4mom.Eta())) ||
-             (2.1 <= fabs(trk4mom.Eta()) && trk4mom.Pt() >= 1.5)));
-  }
-  else if (muonType == (std::string)("TRK") || muonType == (std::string)("TRKSOFT")) {
-    return (fabs(trk4mom.Eta()) < 2.4 &&
-            ((fabs(trk4mom.Eta()) < 1.1 && trk4mom.Pt() >= 3.3) ||
-             (1.1 <= fabs(trk4mom.Eta()) && fabs(trk4mom.Eta()) < 1.3 && trk4mom.Pt() >= 13.2-9.0*fabs(trk4mom.Eta()) ) ||
-             (1.3 <= fabs(trk4mom.Eta()) && trk4mom.Pt() >= 0.8 && trk4mom.Pt() >= 3.02-1.17*fabs(trk4mom.Eta()) )));
-  }
-  else if (muonType == (std::string)("GLBSOFT")) {
-    return (fabs(trk4mom.Eta()) < 2.4 &&
-            ((fabs(trk4mom.Eta()) < 0.3 && trk4mom.Pt() >= 3.4) ||
-             (fabs(trk4mom.Eta()) > 0.3 && fabs(trk4mom.Eta()) < 1.1 && trk4mom.Pt() >= 3.3) ||
-             (fabs(trk4mom.Eta()) > 1.1 && fabs(trk4mom.Eta()) < 1.4 && trk4mom.Pt() >= 7.7-4.0*fabs(trk4mom.Eta()) ) ||
-             (fabs(trk4mom.Eta()) > 1.4 && fabs(trk4mom.Eta()) < 1.55 && trk4mom.Pt() >= 2.1) ||
-             (fabs(trk4mom.Eta()) > 1.55 && fabs(trk4mom.Eta()) < 2.2 && trk4mom.Pt() >= 4.25-1.39*fabs(trk4mom.Eta()) ) ||
-             (fabs(trk4mom.Eta()) > 2.2 && trk4mom.Pt() >= 1.2) ));
-  }  
-  else  std::cout << "ERROR: Incorrect Muon Type" << std::endl;
-
-  return false;
-}
-
-
-bool
-HiOniaAnalyzer::isMuonInAccept(const pat::Muon* aMuon, const std::string muonType) {
-  if (muonType == (std::string)("GLB")) {
-    return (fabs(aMuon->eta()) < 2.4 &&
-            ((fabs(aMuon->eta()) < 1.2 && aMuon->pt() >= 3.5) ||
-             (1.2 <= fabs(aMuon->eta()) && fabs(aMuon->eta()) < 2.1 && aMuon->pt() >= 5.47-1.89*fabs(aMuon->eta())) ||
-             (2.1 <= fabs(aMuon->eta()) && aMuon->pt() >= 1.5)));
-  }
-  else if (muonType == (std::string)("Acceptance2015")) {
-    return (fabs(aMuon->eta()) < 2.4 &&
-            ((fabs(aMuon->eta()) < 1.2 && aMuon->pt() >= 3.5) ||
-             (1.2 <= fabs(aMuon->eta()) && fabs(aMuon->eta()) < 2.1 && aMuon->pt() >= 5.77-1.89*fabs(aMuon->eta())) ||
-             (2.1 <= fabs(aMuon->eta()) && aMuon->pt() >= 1.8)));
-  }
-  else if (muonType == (std::string)("TRK")) { //This is actually softer than the "TRKSOFT" acceptance
-    return (fabs(aMuon->eta()) < 2.4 &&
-            ((fabs(aMuon->eta()) < 0.8 && aMuon->pt() >= 3.3) ||
-             (0.8 <= fabs(aMuon->eta()) && fabs(aMuon->eta()) < 2. && aMuon->p() >= 2.9) ||
-             (2. <= fabs(aMuon->eta()) && aMuon->pt() >= 0.8)));
-  }
-  else if (muonType == (std::string)("GLBSOFT")) {
-    return (fabs(aMuon->eta()) < 2.4 &&
-            ((fabs(aMuon->eta()) < 0.3 && aMuon->pt() >= 3.4) ||
-             (fabs(aMuon->eta()) > 0.3 && fabs(aMuon->eta()) < 1.1 && aMuon->pt() >= 3.3) ||
-             (fabs(aMuon->eta()) > 1.1 && fabs(aMuon->eta()) < 1.4 && aMuon->pt() >= 7.7-4.0*fabs(aMuon->eta()) ) ||
-             (fabs(aMuon->eta()) > 1.4 && fabs(aMuon->eta()) < 1.55 && aMuon->pt() >= 2.1) ||
-             (fabs(aMuon->eta()) > 1.55 && fabs(aMuon->eta()) < 2.2 && aMuon->pt() >= 4.25-1.39*fabs(aMuon->eta()) ) ||
-             (fabs(aMuon->eta()) > 2.2 && aMuon->pt() >= 1.2) ));
-  }
-  else if (muonType == (std::string)("TRKSOFT")) {
-    return (fabs(aMuon->eta()) < 2.4 &&
-	    ((fabs(aMuon->eta()) < 1.1 && aMuon->pt() >= 3.3) ||
-	     (1.1 <= fabs(aMuon->eta()) && fabs(aMuon->eta()) < 1.3 && aMuon->pt() >= 13.2-9.0*fabs(aMuon->eta()) ) ||
-	     (1.3 <= fabs(aMuon->eta()) && aMuon->pt() >= 0.8 && aMuon->pt() >= 3.02-1.17*fabs(aMuon->eta()) )));
-  }
-  else  std::cout << "ERROR: Incorrect Muon Type" << std::endl;
-  
-  return false;
-}
-
-bool
-HiOniaAnalyzer::isSoftMuonBase(const pat::Muon* aMuon) {
-  return (aMuon->isTrackerMuon() &&
-          aMuon->innerTrack()->hitPattern().trackerLayersWithMeasurement() > 5   &&
-          aMuon->innerTrack()->hitPattern().pixelLayersWithMeasurement()   > 0   &&
-          fabs(aMuon->innerTrack()->dxy(RefVtx)) < 0.3 &&
-          fabs(aMuon->innerTrack()->dz(RefVtx)) < 20.
-          );
-}
-
-bool
-HiOniaAnalyzer::isHybridSoftMuon(const pat::Muon* aMuon) {
-  return (isSoftMuonBase(aMuon) &&
-          aMuon->isGlobalMuon()
-          );
-}
-
-bool
-HiOniaAnalyzer::selGlobalMuon(const pat::Muon* aMuon) {
-  
-  if(!aMuon->isGlobalMuon())
-    return false;
-
-  if( _muonSel==(std::string)("GlbTrk") && !aMuon->isTrackerMuon() )
-    return false;
-  
-  if(!_applycuts)
-    return true;
-  
-  bool isInAcc = isMuonInAccept(aMuon, (std::string)(_SofterSgMuAcceptance?"GLBSOFT":"GLB"));
-  bool isGood = (_selTightGlobalMuon ? aMuon->passed(reco::Muon::CutBasedIdTight) : isSoftMuonBase(aMuon) );
-
-  return ( isInAcc && isGood && (!_miniAODcut || PassMiniAODcut(aMuon)) );
-}
-
-bool 
-HiOniaAnalyzer::selTrackerMuon(const pat::Muon* aMuon) {
-  
-  if(!aMuon->isTrackerMuon())
-    return false;
-
-  if(!_applycuts)
-    return true;
-
-  bool isInAcc = isMuonInAccept(aMuon, (std::string)(_SofterSgMuAcceptance?"TRKSOFT":"TRK"));
-  bool isGood = isSoftMuonBase(aMuon);
-
-  return ( isInAcc && isGood && (!_miniAODcut || PassMiniAODcut(aMuon)));
-}
-
-bool
-HiOniaAnalyzer::selGlobalOrTrackerMuon(const pat::Muon* aMuon) {
-
-  if(!aMuon->isGlobalMuon() && !aMuon->isTrackerMuon())
-    return false;
-
-  if(!_applycuts)
-    return true;
-
-  bool isInAcc = isMuonInAccept(aMuon, (std::string)(_SofterSgMuAcceptance?"TRKSOFT":"TRK"));
-  bool isGood = isSoftMuonBase(aMuon);
-
-  return ( isInAcc && isGood && (!_miniAODcut || PassMiniAODcut(aMuon)) );
-}
-
-bool
-HiOniaAnalyzer::selAllMuon(const pat::Muon* aMuon) {
-  return !_miniAODcut || PassMiniAODcut(aMuon);
-}
-
-bool
-HiOniaAnalyzer::PassMiniAODcut(const pat::Muon* aMuon) {
+bool HiOniaAnalyzer::PassMiniAODcut(const pat::Muon* aMuon){
   return aMuon->pt() > 5 || aMuon->isPFMuon() || (aMuon->pt()>1.2 && (aMuon->isGlobalMuon() || aMuon->isStandAloneMuon())) || (aMuon->isTrackerMuon() && aMuon->innerTrack()->quality(reco::TrackBase::highPurity));
-}
+};
 
-bool
-HiOniaAnalyzer::selTrk(const reco::TrackRef aTrk) {
 
-  if(!(aTrk->qualityByName("highPurity") && aTrk->ptError()/aTrk->pt()<0.1))
-    return false;
-
-  if(!_applycuts)
-    return true;
-
-  bool isInAcc = aTrk->pt()>1.2 && fabs(aTrk->eta())<2.4;//(aTrk->pt())>0.2 && fabs(aTrk->eta())<2.4 && aTrk->ptError()/aTrk->pt()<0.1 && fabs(aTrk->dxy(RefVtx))<0.35 && fabs(aTrk->dz(RefVtx))<20; //keep margin in dxy and dz, if the RefVtx is not the good one due to muonlessPV
-
-  return ( isInAcc );
-}
-
-void
-HiOniaAnalyzer::InitEvent()
-{
+void HiOniaAnalyzer::InitEvent(){
   for (unsigned int iTr=1;iTr<NTRIGGERS;++iTr) {
     alreadyFilled[iTr]=false;
   }
@@ -2729,63 +1094,10 @@ HiOniaAnalyzer::InitEvent()
   }
 
   return;
-}
+};
 
-int
-HiOniaAnalyzer::IndexOfThisMuon(TLorentzVector* v1, bool isGen){
-  const auto& mapMuIdx =  (isGen ?  mapGenMuonMomToIndex_ : mapMuonMomToIndex_);
-  const long int& muPt = FloatToIntkey(v1->Pt());
 
-  if (mapMuIdx.count(muPt)==0) return -1;
-  else return mapMuIdx.at(muPt);
-}
-
-int
-HiOniaAnalyzer::IndexOfThisTrack(TLorentzVector* v1, bool isGen){
-  const auto& mapTrkIdx =  (isGen ?  mapTrkMomToIndex_ : mapTrkMomToIndex_);
-  const long int& trkPt = FloatToIntkey(v1->Pt());
-
-  if (mapTrkIdx.count(trkPt)==0) return -1;
-  else return mapTrkIdx.at(trkPt);
-}
-
-int
-HiOniaAnalyzer::IndexOfThisJpsi(int mu1_idx, int mu2_idx, int flipJpsi){
-  int GoodIndex = -1;
-  for(int iJpsi=0; iJpsi<Reco_QQ_size; iJpsi++){
-    if(((Reco_QQ_mumi_idx[iJpsi] == mu1_idx && Reco_QQ_mupl_idx[iJpsi] == mu2_idx) ||
-	(Reco_QQ_mumi_idx[iJpsi] == mu2_idx && Reco_QQ_mupl_idx[iJpsi] == mu1_idx))
-       && flipJpsi == Reco_QQ_flipJpsi[iJpsi]
-       ){
-      GoodIndex = iJpsi;
-      break;
-    }
-  }
-  return GoodIndex;
-}
-
-Short_t 
-HiOniaAnalyzer::MuInSV(TLorentzVector v1, TLorentzVector v2, TLorentzVector v3) {
-
-  //	cout<<"SV collection size : "<<SVs->size()<<endl;
-  int nMuInSV = 0;
-  for(std::vector<reco::Vertex>::const_iterator vt=SVs->begin(); vt!=SVs->end(); ++vt){
-    const reco::Vertex* vtx = &(*vt);
-    int nTrksInSV = 0;
-    for(reco::Vertex::trackRef_iterator it=vtx->tracks_begin();it!=vtx->tracks_end(); ++it) {
-      if((fabs((*it)->pt() - v1.Pt())<1e-3 && fabs((*it)->eta() - v1.Eta())<1e-4) || (fabs((*it)->pt() - v2.Pt())<1e-3 && fabs((*it)->eta() - v2.Eta())<1e-4) || (fabs((*it)->pt() - v3.Pt())<1e-3 && fabs((*it)->eta() - v3.Eta())<1e-4)){
-	nTrksInSV+=1;
-      }
-    }
-    if (nTrksInSV>nMuInSV) nMuInSV=nTrksInSV;
-  }
-  //cout<<"Number of muons from Bc that are in a given SV = "<<nMuInSV<<endl;
-
-  return nMuInSV;
-}
-
-reco::GenParticleRef  
-HiOniaAnalyzer::findDaughterRef(reco::GenParticleRef GenParticleDaughter, int GenParticlePDG) {
+reco::GenParticleRef HiOniaAnalyzer::findDaughterRef(reco::GenParticleRef GenParticleDaughter, int GenParticlePDG){
 
   reco::GenParticleRef GenParticleTmp = GenParticleDaughter;
   bool foundFirstDaughter = false;
@@ -2813,33 +1125,10 @@ HiOniaAnalyzer::findDaughterRef(reco::GenParticleRef GenParticleDaughter, int Ge
 
   return GenParticleDaughter;
 
-}
+};
 
-std::pair<bool, reco::GenParticleRef>
-HiOniaAnalyzer::findBcMotherRef(reco::GenParticleRef GenParticleMother, int BcPDG) {
 
-  bool FoundBc = false;
-  for(int i=0; i<1000; ++i) {
-    if ((fabs(GenParticleMother->pdgId()) == BcPDG) && (GenParticleMother->status() == 2) && (GenParticleMother->numberOfDaughters() >= 2)) {
-      FoundBc = true;
-      break;
-    }
-    else if (GenParticleMother.isNonnull() && GenParticleMother->numberOfMothers()>0){
-      GenParticleMother = GenParticleMother->motherRef();
-    }
-    else {
-      break;}
-  }
-
-  std::pair<bool, reco::GenParticleRef> res;
-  res.first = FoundBc; res.second = GenParticleMother;
-  return res;
-
-}
-
-void
-HiOniaAnalyzer::fillGenInfo()
-{
+void HiOniaAnalyzer::fillGenInfo(){
   if (Gen_QQ_size >= Max_QQ_size) {
     std::cout << "Too many dimuons: " << Gen_QQ_size << std::endl;
     std::cout << "Maximum allowed: " << Max_QQ_size << std::endl;
@@ -3018,12 +1307,10 @@ HiOniaAnalyzer::fillGenInfo()
     }
   }  
   return;
-}
+};
 
 //Find the indices of the reconstructed muon matching each generated muon, and vice versa
-void
-HiOniaAnalyzer::fillMuMatchingInfo()
-{
+void HiOniaAnalyzer::fillMuMatchingInfo(){
 
   //initialize Gen_mu_whichRec
   for (int igen=0;igen<Gen_mu_size;igen++){
@@ -3048,11 +1335,10 @@ HiOniaAnalyzer::fillMuMatchingInfo()
     if(foundGen>-1) Gen_mu_whichRec[foundGen] = irec;
   }
   
-}
+};
 
 //Find the indices of the reconstructed J/psi matching each generated J/psi (when the two daughter muons are reconstructed), and vice versa
-void
-HiOniaAnalyzer::fillQQMatchingInfo(){
+void HiOniaAnalyzer::fillQQMatchingInfo(){
   for (int igen=0;igen<Gen_QQ_size;igen++){
     Gen_QQ_whichRec[igen] = -1;
     int Reco_mupl_idx = Gen_mu_whichRec[Gen_QQ_mupl_idx[igen]]; //index of the reconstructed mupl associated to the generated mupl of Jpsi
@@ -3083,149 +1369,10 @@ HiOniaAnalyzer::fillQQMatchingInfo(){
     }      
   }
   
-}
-
-//Build the visible Bc when the three daugther muons are reconstructed. Record indices between Gen and Rec 
-void
-HiOniaAnalyzer::fillBcMatchingInfo(){
-  for (int igen=0;igen<Gen_Bc_size;igen++){
-    Gen_3mu_whichRec[igen] = -1;
-
-    //Build the visible Gen Bc (sum of lorentzvectors of the three gen muons)
-    TLorentzVector gen_3mu_4mom = *((TLorentzVector*)Gen_QQ_4mom->ConstructedAt(Gen_Bc_QQ_idx[igen]))
-                                  + *((TLorentzVector*)Gen_mu_4mom->ConstructedAt(Gen_Bc_muW_idx[igen]));
-    new((*Gen_3mu_4mom)[igen])TLorentzVector(gen_3mu_4mom);
-    Gen_3mu_4mom_pt.push_back(gen_3mu_4mom.Pt());
-    Gen_3mu_4mom_eta.push_back(gen_3mu_4mom.Eta());
-    Gen_3mu_4mom_phi.push_back(gen_3mu_4mom.Phi());
-    Gen_3mu_4mom_m.push_back(gen_3mu_4mom.M());
-
-    //Look for index of reconstructed muon
-    int Reco_muW_idx = Gen_mu_whichRec[Gen_Bc_muW_idx[igen]]; //index of the reconstructed muW associated to the generated muW of Bc
-    int Reco_mupl_idx = Gen_mu_whichRec[Gen_QQ_mupl_idx[Gen_Bc_QQ_idx[igen]]]; //index of the reconstructed mupl associated to the generated mupl of Jpsi
-    int Reco_mumi_idx = Gen_mu_whichRec[Gen_QQ_mumi_idx[Gen_Bc_QQ_idx[igen]]]; //index of the reconstructed mumi associated to the generated mumi of Jpsi
-
-    if((Reco_mupl_idx>=0) && (Reco_mumi_idx>=0) && (Reco_muW_idx>=0)){   //Search for Reco_Bc only if the three muons are reco
-
-      for (int irec=0;irec<Reco_3mu_size;irec++){
-        if( ((Reco_mupl_idx == Reco_3mu_mupl_idx[irec]) && (Reco_mumi_idx == Reco_3mu_mumi_idx[irec]) && (Reco_muW_idx == Reco_3mu_muW_idx[irec])) ||    //the charges might be wrong in reco 
-            ((Reco_mupl_idx == Reco_3mu_mupl_idx[irec]) && (Reco_mumi_idx == Reco_3mu_muW_idx[irec]) && (Reco_muW_idx == Reco_3mu_mumi_idx[irec])) ||
-            ((Reco_mupl_idx == Reco_3mu_muW_idx[irec]) && (Reco_mumi_idx == Reco_3mu_mumi_idx[irec]) && (Reco_muW_idx == Reco_3mu_mupl_idx[irec])) ||
-            ((Reco_mupl_idx == Reco_3mu_muW_idx[irec]) && (Reco_mumi_idx == Reco_3mu_mupl_idx[irec]) && (Reco_muW_idx == Reco_3mu_mumi_idx[irec])) ||
-            ((Reco_mupl_idx == Reco_3mu_mumi_idx[irec]) && (Reco_mumi_idx == Reco_3mu_muW_idx[irec]) && (Reco_muW_idx == Reco_3mu_mupl_idx[irec])) ||
-            ((Reco_mupl_idx == Reco_3mu_mumi_idx[irec]) && (Reco_mumi_idx == Reco_3mu_mupl_idx[irec]) && (Reco_muW_idx == Reco_3mu_muW_idx[irec]))
-            ){
-          Gen_3mu_whichRec[igen] = irec;
-          break;
-	}
-      }
- 
-      if((Gen_3mu_whichRec[igen]==-1) ){
-	Gen_3mu_whichRec[igen]=-2; //The three muons were reconstructed, but the associated trimuon was not selected
-      }
-    }
-
-  }
-
-  for (int irec=0;irec<Reco_3mu_size;irec++){
-
-    //Find the index of generated Bc associated to a reco trimuon 
-    Reco_3mu_whichGen[irec] = -1;
-    for (int igen=0;igen<Gen_Bc_size;igen++){
-      if((Gen_3mu_whichRec[igen] == irec)){
-	Reco_3mu_whichGen[irec] = igen;
-	break;
-      }
-    }
-
-    //Match the muW to a gen particle, and find out if it comes from the B meson hard process
-    if(_genealogyInfo){
-
-      //cout<<"Checking genealogy for Reco_3mu #"<<irec<<endl;
-      //cout<<"Reco_muW simExtType,simPdgId,simMotherPdgId = "<<Reco_mu_simExtType[Reco_3mu_muW_idx[irec]]<<" "<<Reco_mu_simPdgId[Reco_3mu_muW_idx[irec]]<<" "<<Reco_mu_simMotherPdgId[Reco_3mu_muW_idx[irec]]<<endl;
-      Reco_3mu_muW_isGenJpsiBro[irec] = false;
-      Reco_3mu_muW_trueId[irec] = 0;
-      
-      int goodRecQQidx = (Reco_3mu_QQ1_idx[irec]>-1)?(Reco_3mu_QQ1_idx[irec]):(Reco_3mu_QQ2_idx[irec]);
-      if(Reco_3mu_QQ1_idx[irec]>-1 && Reco_3mu_QQ2_idx[irec]>-1){ //If both OS pairs are reconstructed, choose the reco QQ that is matched to a gen
-	goodRecQQidx = (Reco_QQ_whichGen[Reco_3mu_QQ1_idx[irec]]>-1)?(Reco_3mu_QQ1_idx[irec]):(Reco_3mu_QQ2_idx[irec]);}
-      int genQQidx = Reco_QQ_whichGen[goodRecQQidx];
-      int newmuWidx = (goodRecQQidx==Reco_3mu_QQ1_idx[irec])?(Reco_3mu_muW_idx[irec]):(Reco_3mu_muW2_idx[irec]); //use the muW_idx corresponding to the chosen QQ dimuon
-      int genMuWidx = Reco_mu_whichGen[newmuWidx];
-
-      //If the muon is fake, match it to whatever generated particles 
-      if(genMuWidx==-1){
-	TLorentzVector* recmuW = (TLorentzVector*)Reco_mu_4mom->ConstructedAt(newmuWidx);
-	bool SureDecayInFlight = (fabs(Reco_mu_simExtType[newmuWidx]) == 4);
-	bool Unmatched = (fabs(Reco_mu_simExtType[newmuWidx]) == 0);
-	float dRmax = SureDecayInFlight?0.3:(Unmatched?0.15:0.1);
-	float dRmin = dRmax;
-	float matchedPhi = 5;
-	
-	//Loop over gen particles
-	//cout<<"Looking for a (charged track) gen particle matching the fake Reco_muW of pt,eta = "<<recmuW->Pt()<<" "<<recmuW->Eta()<<endl;
-	for(std::vector<reco::GenParticle>::const_iterator it=collGenParticles->begin();
-	    it!=collGenParticles->end();++it) {
-	  const reco::GenParticle* gen = &(*it);
-
-	  if (isChargedTrack(gen->pdgId()) //&& gen->status() == 1
-	      ) {
-	    TLorentzVector genP = lorentzMomentum(gen->p4());
-	    float dR = recmuW->DeltaR(genP);
-	    if(dR<dRmin
-	       && (!Unmatched || (genP.Pt()>0.6*recmuW->Pt()) ) // if unmatched, probably a ghost, so at least have a decent Pt agreement (leaving possibility for energy loss of a decayInFlight) for the particle causing the ghost
-	       ){
-	      dRmin = dR;
-	      Reco_3mu_muW_trueId[irec] = gen->pdgId();
-	      matchedPhi = genP.Phi();
-	      //cout<<"   ........Found matched gen particle for fake reco muon: ID, pt, eta, deltaR = "<<Reco_3mu_muW_trueId[irec]<<" "<<genP.Pt()<<" "<<genP.Eta()<<" "<<dRmin<<endl;
-	    }
-	  }
-	}
-	//end loop over gen particles	
-
-	//If the Jpsi is true, check if the matched gen particle was part of the B-parent process
-	if(genQQidx>-1){
-	  for(auto&& bro: _Gen_QQ_MomAndTrkBro[genQQidx] ){
-	    if(fabs( bro->phi() - matchedPhi ) < 1e-6 && isChargedTrack(bro->pdgId())){
-	      Reco_3mu_muW_isGenJpsiBro[irec] = true;
-	      //cout<<"   !!!!!!!!!!!!    FAKE muW is from B parent"<<endl;
-	      break;
-	    }
-	  }
-	}
-
-      }
-
-      //If the muon is true, take the associated gen muon, and find if it comes from the B parent of Jpsi
-      if(genMuWidx>-1){
-	Reco_3mu_muW_trueId[irec] = (Gen_mu_charge[genMuWidx]==-1)?13:(-13);
-
-	if(genQQidx>-1){
-	  //cout<<"Checking if this gen muon comes from B parent: pt,eta = "<<((TLorentzVector*)Gen_mu_4mom->ConstructedAt(genMuWidx))->Pt()<<" "<<((TLorentzVector*)Gen_mu_4mom->ConstructedAt(genMuWidx))->Eta()<<endl;
-	  for(auto&& bro: _Gen_QQ_MomAndTrkBro[genQQidx] ){
-	    if(fabs(bro->pdgId())==13 && fabs( bro->phi() - ((TLorentzVector*)Gen_mu_4mom->ConstructedAt(genMuWidx))->Phi() ) < 1e-6){
-	      Reco_3mu_muW_isGenJpsiBro[irec] = true;
-	      //cout<<"   !!!!!!!!!!!!    True muW is from B parent: pt, eta = "<<((TLorentzVector*)Gen_mu_4mom->ConstructedAt(genMuWidx))->Pt()<<" "<<((TLorentzVector*)Gen_mu_4mom->ConstructedAt(genMuWidx))->Eta()<<endl;
-	      break;
-	    }
-	  }
-	}
-	else {//cout<<"True Muon but fake Jpsi"<<endl;
-	}
-      }
-     
-      //cout<<endl; 
-    }
-
-  }
+};
 
 
-}
-
-void
-HiOniaAnalyzer::fillRecoTracks()
-{
+void HiOniaAnalyzer::fillRecoTracks(){
   if (collTracks.isValid()) {
     for(unsigned int tidx=0; tidx<collTracks->size();tidx++) {
       const reco::TrackRef track(collTracks, tidx);
@@ -3299,11 +1446,9 @@ HiOniaAnalyzer::fillRecoTracks()
   }
   
   return;
-}
+};
 
-void
-HiOniaAnalyzer::fillRecoMuons(int iCent)
-{
+void HiOniaAnalyzer::fillRecoMuons(int iCent){
   int nL1DoubleMu0Muons=0;
   int nGoodMuons=0;
   int nGoodMuonsNoTrig=0;
@@ -3433,11 +1578,9 @@ HiOniaAnalyzer::fillRecoMuons(int iCent)
   hL1DoubleMu0->Fill(nL1DoubleMu0Muons);
 
   return;
-}
+};
 
-void
-HiOniaAnalyzer::InitTree()
-{
+void HiOniaAnalyzer::InitTree(){
 
   Reco_mu_4mom = new TClonesArray("TLorentzVector", Max_mu_size);
   Reco_mu_L1_4mom = new TClonesArray("TLorentzVector", Max_mu_size);
@@ -3795,9 +1938,7 @@ HiOniaAnalyzer::InitTree()
 }
 
 // ------------ method called once each job just before starting event loop  ------------
-void 
-HiOniaAnalyzer::beginJob()
-{
+void  HiOniaAnalyzer::beginJob(){
   //fOut = new TFile(_histfilename.c_str(), "RECREATE");
   InitTree();
 
@@ -3896,10 +2037,9 @@ HiOniaAnalyzer::beginJob()
   hZVtx->Sumw2();
 
   return;
-}
+};
 
-void 
-HiOniaAnalyzer::beginRun(const edm::Run& iRun, const edm::EventSetup& iSetup) {
+void HiOniaAnalyzer::beginRun(const edm::Run& iRun, const edm::EventSetup& iSetup){
   //init HLTConfigProvider
   
   EDConsumerBase::Labels labelTriggerResults;
@@ -3938,369 +2078,14 @@ HiOniaAnalyzer::beginRun(const edm::Run& iRun, const edm::EventSetup& iSetup) {
   }
 
   return;
-}
+};
 
 // ------------ method called once each job just after ending the event loop  ------------
-void 
-HiOniaAnalyzer::endJob() {
+void  HiOniaAnalyzer::endJob(){
   std::cout << "Total number of events = " << nEvents << std::endl;
   std::cout << "Total number of passed candidates = " << passedCandidates << std::endl;
-  /*
-  fOut->cd();
-  hStats->Write();
-  hCent->Write();
-  hPileUp->Write();
-  hZVtx->Write();
-
-  if (_fillTree)  
-    myTree->Write();
-
-  hGoodMuonsNoTrig->Write();
-  hGoodMuons->Write();
-  hL1DoubleMu0->Write();
-
-  if (_fillHistos) {
-    // muons
-    if (_combineCategories)
-      myRecoGlbMuonHistos->Write(fOut);
-    else {
-      myRecoGlbMuonHistos->Write(fOut);
-      myRecoTrkMuonHistos->Write(fOut);
-    }
-
-    // J/psi
-    if (_combineCategories)
-      myRecoJpsiHistos->Write(fOut);
-    else {
-      myRecoJpsiGlbGlbHistos->Write(fOut);
-      myRecoJpsiGlbTrkHistos->Write(fOut);
-      myRecoJpsiTrkTrkHistos->Write(fOut);
-    }
-  }
-  */
   return;
-}
+};
 
-TLorentzVector
-HiOniaAnalyzer::lorentzMomentum(const reco::Candidate::LorentzVector& p) {
-  TLorentzVector res;
-  res.SetPtEtaPhiM(p.pt(), p.eta(), p.phi(), p.mass());
 
-  return res;
-}
 
-void
-HiOniaAnalyzer::hltReport(const edm::Event &iEvent ,const edm::EventSetup& iSetup)
-{
-  std::map<std::string, bool> mapTriggernameToTriggerFired;
-  std::map<std::string, unsigned int> mapTriggernameToHLTbit;
-
-  for(std::vector<std::string>::const_iterator it=theTriggerNames.begin(); it !=theTriggerNames.end(); ++it){
-    mapTriggernameToTriggerFired[*it]=false;
-    mapTriggernameToHLTbit[*it]=1000;
-  }
-
-  // HLTConfigProvider
-  if ( hltConfigInit ) {
-    //! Use HLTConfigProvider
-    const unsigned int n= hltConfig.size();
-    for (std::map<std::string, unsigned int>::iterator it = mapTriggernameToHLTbit.begin(); it != mapTriggernameToHLTbit.end(); it++) {
-      unsigned int triggerIndex= hltConfig.triggerIndex( triggerNameMap.at(it->first) );
-      if (it->first == "NoTrigger") continue;
-      if (triggerIndex >= n) {
-	if (_checkTrigNames) std::cout << "[HiOniaAnalyzer::hltReport] --- TriggerName " << it->first << " not available in config!" << std::endl;
-      }
-      else {
-        it->second= triggerIndex;
-        //      std::cout << "[HiOniaAnalyzer::hltReport] --- TriggerName " << it->first << " available in config!" << std::endl;
-      }
-    }
-  }
-    
-  // Get Trigger Results
-  iEvent.getByToken( _tagTriggerResultsToken, collTriggerResults );
-  if ( collTriggerResults.isValid() && (collTriggerResults->size()==hltConfig.size()) ){
-    //    std::cout << "[HiOniaAnalyzer::hltReport] --- J/psi TriggerResults IS valid in current event" << std::endl;
-      
-    // loop over Trigger Results to check if paths was fired
-    for(std::vector< std::string >::iterator itHLTNames= theTriggerNames.begin(); itHLTNames != theTriggerNames.end(); itHLTNames++){
-      const std::string triggerPathName =  *itHLTNames;
-      if ( mapTriggernameToHLTbit[triggerPathName] < 1000 ) {
-        if (collTriggerResults->accept( mapTriggernameToHLTbit[triggerPathName] ) ){
-          mapTriggerNameToIntFired_[triggerPathName] = 3;
-        }
-        if (_isMC) {
-          mapTriggerNameToPrescaleFac_[triggerPathName] = 1;
-        } else {
-          //-------prescale factor------------
-          if ( hltPrescaleInit && hltPrescaleProvider.prescaleSet(iEvent,iSetup)>=0 ) {
-            auto const detailedPrescaleInfo = hltPrescaleProvider.prescaleValuesInDetail<double,double>(iEvent, iSetup, triggerNameMap.at(triggerPathName));
-            //std::pair<std::vector<std::pair<std::string,int> >,int> detailedPrescaleInfo = hltPrescaleProvider.prescaleValuesInDetail(iEvent, iSetup, triggerPathName);
-            //get HLT prescale info from hltPrescaleProvider     
-            const int hltPrescale = detailedPrescaleInfo.second;
-            //get L1 prescale info from hltPrescaleProvider
-            int l1Prescale = -1;
-            if (detailedPrescaleInfo.first.size()==1) {
-              l1Prescale = detailedPrescaleInfo.first.at(0).second;
-            }
-            else if (detailedPrescaleInfo.first.size()>1) {
-              l1Prescale = 1; // Means it is a complex l1 seed, and for us it is only Mu3 OR Mu5, both of them unprescaled at L1
-            }
-            else if(_checkTrigNames) {
-              std::cout << "[HiOniaAnalyzer::hltReport] --- L1 prescale was NOT found for TriggerName " << triggerPathName  << " , default L1 prescale value set to 1 " <<  std::endl;
-            }
-            //compute the total prescale = HLT prescale * L1 prescale
-            mapTriggerNameToPrescaleFac_[triggerPathName] = hltPrescale * l1Prescale;
-          }
-        }
-      }
-    }
-  } else std::cout << "[HiOniaAnalyzer::hltReport] --- TriggerResults NOT valid in current event" << std::endl;
-
-  return;
-}
-
-bool 
-HiOniaAnalyzer::isAbHadron(int pdgID) {
-  return (abs(pdgID) == 511 || abs(pdgID) == 521 || abs(pdgID) == 531 || abs(pdgID) == 5122 || abs(pdgID) == 541);
-
-}
-
-bool
-HiOniaAnalyzer::isNeutrino(int pdgID) {
-  return (abs(pdgID) == 14 || abs(pdgID) == 16 || abs(pdgID) == 18);
-}
-
-bool 
-HiOniaAnalyzer::isAMixedbHadron(int pdgID, int momPdgID) {
-
-  if ((abs(pdgID) == 511 && abs(momPdgID) == 511 && pdgID*momPdgID < 0) || 
-      (abs(pdgID) == 531 && abs(momPdgID) == 531 && pdgID*momPdgID < 0)) 
-      return true;
-  return false;
-
-}
-
-reco::GenParticleRef   
-HiOniaAnalyzer::findMotherRef(reco::GenParticleRef GenParticleMother, int GenParticlePDG) {
-
-  for(int i=0; i<1000; ++i) {
-    if (GenParticleMother.isNonnull() && (GenParticleMother->pdgId()==GenParticlePDG) && GenParticleMother->numberOfMothers()>0) {        
-      GenParticleMother = GenParticleMother->motherRef();
-    } else break;
-  }
-  return GenParticleMother;
-
-}
-
-bool HiOniaAnalyzer::isChargedTrack(int pdgId){
-  return ((fabs(pdgId) == 211) || (fabs(pdgId) == 321) || (fabs(pdgId) == 2212) || (fabs(pdgId) == 11) || (fabs(pdgId) == 13));
-}
-
-std::vector<reco::GenParticleRef> HiOniaAnalyzer::GenBrothers(reco::GenParticleRef GenParticleMother, int GenJpsiPDG){
-  bool foundJpsi = false;
-  std::vector<reco::GenParticleRef> res;
-
-  if(!GenParticleMother.isNonnull()) return res;
-  //if(Reco_3mu_size>0) cout<<"\nScanning daughters of Jpsi "<<GenJpsiPDG<<" mother, pdg = "<<GenParticleMother->pdgId()<<endl; 
-  for(int i=0;i<(int)GenParticleMother->numberOfDaughters();i++){    
-    reco::GenParticleRef dau = findDaughterRef(GenParticleMother->daughterRef(i), GenParticleMother->pdgId());
-    for(int l=0;l<100;l++){ //avoid having a daughter of same pdgId
-      if(!(dau.isNonnull() && dau->status()>0 && dau->status()<1000)) break;
-      if(dau->pdgId()==GenParticleMother->pdgId() && dau->numberOfDaughters()==1)
-	dau = findDaughterRef(dau->daughterRef(0), dau->pdgId());
-      else break;
-    }
-
-    if(!(dau.isNonnull() && dau->status()>0 && dau->status()<1000) ) continue;
-    //if(Reco_3mu_size>0) cout<<"Daughter #"<<i<<" pdg = "<< dau->pdgId()<<" pt,eta = "<<dau->pt()<<" "<<dau->eta()<<endl;
-    if(isChargedTrack(dau->pdgId())){
-      res.push_back(dau);}
-    if(dau->pdgId()==GenJpsiPDG) {
-      foundJpsi = true; //continue;
-    }
-
-    for(int j=0;j<(int)dau->numberOfDaughters();j++){
-      reco::GenParticleRef grandDau = findDaughterRef(dau->daughterRef(j), dau->pdgId());
-      for(int l=0;l<100;l++){ //avoid having a daughter of same pdgId
-	if(!(grandDau.isNonnull() && grandDau->status()>0 && grandDau->status()<1000)) break;
-	if(grandDau->pdgId()==dau->pdgId() && grandDau->numberOfDaughters()==1)
-	  grandDau = findDaughterRef(grandDau->daughterRef(0), grandDau->pdgId());
-	else break;
-      }
-
-      if(!(grandDau.isNonnull() && grandDau->status()>0 && grandDau->status()<1000)) continue;
-      //if(Reco_3mu_size>0) cout<<"    grand-daughter #"<<j<<" pdg = "<< grandDau->pdgId()<<" pt,eta = "<<grandDau->pt()<<" "<<grandDau->eta()<<endl;
-      if(isChargedTrack(grandDau->pdgId())){
-	res.push_back(grandDau);}
-      if(grandDau->pdgId()==GenJpsiPDG) {
-	foundJpsi = true; //continue;
-      }
-
-      for(int k=0;k<(int)grandDau->numberOfDaughters();k++){
-	reco::GenParticleRef ggrandDau = findDaughterRef(grandDau->daughterRef(k), grandDau->pdgId());
-	for(int l=0;l<100;l++){ //avoid having a daughter of same pdgId
-	  if(!(ggrandDau.isNonnull() && ggrandDau->status()>0 && ggrandDau->status()<1000)) break;
-	  if(ggrandDau->pdgId()==grandDau->pdgId() && ggrandDau->numberOfDaughters()==1)
-	    ggrandDau = findDaughterRef(ggrandDau->daughterRef(0), ggrandDau->pdgId());
-	  else break;
-	}
-
-	if(!(ggrandDau.isNonnull() && ggrandDau->status()>0 && ggrandDau->status()<1000)) continue;
-	//if(Reco_3mu_size>0) cout<<"        grand-grand-daughter #"<<k<<" pdg = "<< ggrandDau->pdgId()<<" pt,eta = "<<ggrandDau->pt()<<" "<<ggrandDau->eta()<<endl;
-	if(isChargedTrack(ggrandDau->pdgId())){
-	  res.push_back(ggrandDau);}
-	if(ggrandDau->pdgId()==GenJpsiPDG) {
-	  foundJpsi = true; //continue;
-	}
-      }
-    }
-  }
-
-  if(!foundJpsi){
-    cout<<"!!!!!!!!!!!!!!!!!!!!!!!!!!!! Incoherence in genealogy: Jpsi not found in the daughters!\n"<<endl;
-  }
-  // if(!isAbHadron(GenParticleMother->pdgId())){
-  //   cout<<"\n!!!!!!!!!!!!!!!!!!!!!!!!!!!! Jpsi ancestor is not a b-hadron! pdgID(mother of this ancestor) = "<<findMotherRef(GenParticleMother->motherRef() , GenParticleMother->pdgId())->pdgId()<<endl;
-  // }  
-
-  return res;
-}
-
-std::pair< std::vector<reco::GenParticleRef>, std::pair<float, float> >  
-HiOniaAnalyzer::findGenMCInfo(const reco::GenParticle* genJpsi) {
-
-  float trueLife = -99.;
-  float trueLife3D = -99.;
-  std::vector<reco::GenParticleRef> JpsiBrothers;
-
-  if (genJpsi->numberOfMothers()>0) {
-    TVector3 trueVtx(0.0,0.0,0.0);
-    TVector3 trueP(0.0,0.0,0.0);
-    TVector3 trueVtxMom(0.0,0.0,0.0);
-
-    trueVtx.SetXYZ(genJpsi->vertex().x(),genJpsi->vertex().y(),genJpsi->vertex().z());
-    trueP.SetXYZ(genJpsi->momentum().x(),genJpsi->momentum().y(),genJpsi->momentum().z());
-
-    bool aBhadron = false;
-    reco::GenParticleRef Jpsimom_final;
-    reco::GenParticleRef Jpsimom = findMotherRef(genJpsi->motherRef(), genJpsi->pdgId());      
-
-    if (Jpsimom.isNull()) {
-      std::pair<float, float> trueLifePair = std::make_pair(trueLife, trueLife3D);
-      std::pair< std::vector<reco::GenParticleRef>, std::pair<float, float> > result = std::make_pair(JpsiBrothers, trueLifePair);
-      return result;
-    } 
-    else if (Jpsimom->numberOfMothers()<=0) {
-      if (isAbHadron(Jpsimom->pdgId())) {  
-	Jpsimom_final = Jpsimom;
-	aBhadron = true;
-      }
-    } 
-
-    else {
-      reco::GenParticleRef Jpsigrandmom = findMotherRef(Jpsimom->motherRef(), Jpsimom->pdgId());   
-      if (isAbHadron(Jpsimom->pdgId())) {
-        if (Jpsigrandmom.isNonnull() && isAMixedbHadron(Jpsimom->pdgId(),Jpsigrandmom->pdgId())) {       
-	  Jpsimom_final = Jpsigrandmom;
-        } 
-        else {                  
-	Jpsimom_final = Jpsimom;
-        }
-        aBhadron = true;
-      } 
-
-      else if (Jpsigrandmom.isNonnull() && isAbHadron(Jpsigrandmom->pdgId()))  {  
-        if (Jpsigrandmom->numberOfMothers()<=0) {
-	  Jpsimom_final = Jpsigrandmom;
-        } 
-        else { 
-          reco::GenParticleRef JpsiGrandgrandmom = findMotherRef(Jpsigrandmom->motherRef(), Jpsigrandmom->pdgId());
-          if (JpsiGrandgrandmom.isNonnull() && isAMixedbHadron(Jpsigrandmom->pdgId(),JpsiGrandgrandmom->pdgId())) {
-	    Jpsimom_final = JpsiGrandgrandmom;
-          } 
-          else {
-	    Jpsimom_final = Jpsigrandmom;
-          }
-        }
-        aBhadron = true;
-      }
-
-      //This is to forcefully find the b-like mother of Jpsi 
-      else if (Jpsigrandmom.isNonnull() && Jpsigrandmom->numberOfMothers()>0){
-	reco::GenParticleRef JpsiGrandgrandmom = findMotherRef(Jpsigrandmom->motherRef(), Jpsigrandmom->pdgId());
-	if(JpsiGrandgrandmom.isNonnull() && isAbHadron(JpsiGrandgrandmom->pdgId())){
-          Jpsimom_final = JpsiGrandgrandmom;
-	  aBhadron = true;
-	}
-      }
-
-    }
-    if (!aBhadron) {
-      Jpsimom_final = Jpsimom;
-    }
-
-    if (Jpsimom_final.isNonnull()){
-      trueVtxMom.SetXYZ(Jpsimom_final->vertex().x(),Jpsimom_final->vertex().y(),Jpsimom_final->vertex().z());
-      if(_genealogyInfo && Reco_3mu_size>0
-	 ){
-	JpsiBrothers = GenBrothers(Jpsimom_final, genJpsi->pdgId() );
-      }
-      JpsiBrothers.insert(JpsiBrothers.begin(), Jpsimom_final);
-    }
-    
-    TVector3 vdiff = trueVtx - trueVtxMom;
-    trueLife = vdiff.Perp()*JpsiPDGMass/trueP.Perp();
-    trueLife3D = vdiff.Mag()*JpsiPDGMass/trueP.Mag();
-
-  }
-
-  std::pair<float, float> trueLifePair = std::make_pair(trueLife, trueLife3D);
-  std::pair<std::vector<reco::GenParticleRef>, std::pair<float, float> > result = std::make_pair(JpsiBrothers, trueLifePair);
-  return result;
-
-}
-
-std::pair<int, std::pair<float, float> >
-HiOniaAnalyzer::findGenBcInfo(reco::GenParticleRef genBc, const reco::GenParticle* genJpsi) {
-
-  int momBcID = 0;
-  float trueLife = -99.;
-
-  TVector3 trueVtx(0.0,0.0,0.0);
-  TVector3 trueP(0.0,0.0,0.0);
-  TVector3 trueVtxMom(0.0,0.0,0.0);
-
-  trueVtx.SetXYZ(genJpsi->vertex().x(),genJpsi->vertex().y(),genJpsi->vertex().z());
-  trueVtxMom.SetXYZ(genBc->vertex().x(),genBc->vertex().y(),genBc->vertex().z());
-  trueP.SetXYZ(genBc->momentum().x(),genBc->momentum().y(),genBc->momentum().z());
-
-  TVector3 vdiff = trueVtx - trueVtxMom;
-  trueLife = vdiff.Perp()*BcPDGMass/trueP.Perp();
-
-  std::pair<float, float> trueLifePair = std::make_pair(trueLife, trueLife);
-  std::pair<int, std::pair<float, float> > result = std::make_pair(momBcID, trueLifePair);
-  return result;
-
-}
-
-int HiOniaAnalyzer::muonIDmask(const pat::Muon* muon)
-{
-   int mask = 0;
-   int type;
-   for (type=muon::All; type<=muon::RPCMuLoose; type++)
-      if (muon->hasUserInt(Form("muonID_%d", type)) ? muon->userInt(Form("muonID_%d", type)) : muon::isGoodMuon(*muon, muon::SelectionType(type)))
-         mask = mask | (int) pow(2, type);
-
-   return mask;
-}
-
-long int HiOniaAnalyzer::FloatToIntkey(float v)
-{
-  float vres = fabs(v);
-  while(vres>0.1) vres = vres/10; //Assume argument v is always above 0.1, true for abs(Pt)
-  return (long int) (10000000*vres); // Precision 10^-6 (i.e. 7-1) on the comparison
-}
-
-//define this as a plug-in
-DEFINE_FWK_MODULE(HiOniaAnalyzer);
